refactor: extract pure template/cache helpers for direct testability
Splits the Obsidian-coupled I/O shells from the pure logic they depend
on, so tests no longer need to re-implement the pure pieces.

* New `src/templateRender.ts` — `applyTemplate` and its helpers, zero
  Obsidian imports. Hoists the granularity-token and weekday regex
  patterns to module-level constants (they were rebuilt on every call).
* `canonicalKey` moves from `cache.ts` to `cacheSearch.ts`, the natural
  home for pure cache helpers.
* `template.test.ts` and `cache.test.ts` now import the real functions
  instead of re-declaring them — deletes ~140 lines of drift-prone
  duplicates.
* `template.ts` keeps the I/O layer (readTemplate, applyTemplateToFile,
  getNoteCreationPath); `cache.ts` keeps the vault-event orchestration.

CLAUDE.md and THEORY.md updated to describe the new module layout.

Closes #116, #93, #115.
