+++ main.js
@@ -1,27 +1,27 @@
-let e=require(`obsidian`);function t(e,t,n){let r=0,i=e.length-1;for(;r<=i;){let a=r+i>>>1,o=e[a];if(o===t)return e[a+(n===`forwards`?1:-1)]??null;o<t?r=a+1:i=a-1}return null}var n={day:`YYYY-MM-DD`,week:`gggg-[W]ww`,month:`YYYY-MM`,year:`YYYY`},r={enabled:!1,format:``,folder:``,templatePath:void 0},i={granularities:{day:{...r},week:{...r},month:{...r},year:{...r}}},a=[`sunday`,`monday`,`tuesday`,`wednesday`,`thursday`,`friday`,`saturday`],o=`calendar`,s=Symbol(`displayedMonth`),c=[`day`,`week`,`month`,`year`];function l(e,t){return e.granularities[t].format||n[t]}function u(e,t){let r=e.granularities[t].format;if(!r)return[n[t]];let i=/[^/]*$/.exec(r);return i?[r,i[0]]:[r]}function d(e,t){return e.granularities[t]}function f(e){return c.filter(t=>e.granularities[t].enabled)}function p(e){return e.replace(/\[[^\]]*\]/g,``).replace(/\\./g,``)}function m(e){return e.indexOf(`/`)===-1?e:e.split(`/`).pop()??``}function h(e){return!/[?<>\\:*|"]/g.test(e)&&!/[\x00-\x1f\x80-\x9f]/g.test(e)&&!/^\.+$/.test(e)&&!/^(con|prn|aux|nul|com[0-9]|lpt[0-9])(\..*)?$/i.test(e)}function g(e,t){if(!e)return``;if(!h(e))return`Format contains illegal characters`;if(t===`day`){let t=window.moment().format(e);if(!window.moment(t,e,!0).isValid())return`Failed to parse format`}return``}function _(e){let t=m(e).replace(/\[[^\]]*\]/g,``);return![`M`,`D`].every(e=>t.includes(e))||!(t.includes(`Y`)||t.includes(`y`))}function v(e,t){let n=window.moment().format(e);return window.moment(n,e,!0).isValid()?p(e).includes(`/`)&&t===`day`&&_(e)?`fragile-basename`:`valid`:`loose-parsing`}function y(...e){let t=[];for(let n=0,r=e.length;n<r;n++)t=t.concat(e[n].split(`/`));let n=[];for(let e=0,r=t.length;e<r;e++){let r=t[e];!r||r===`.`||n.push(r)}return t[0]===``&&n.unshift(``),n.join(`/`)}function b(){let{moment:e}=window,t=e.localeData().firstDayOfWeek(),n=[...a];for(;t;){let e=n.shift();e&&n.push(e),t--}return n}function x(e){let t=b().indexOf(e.toLowerCase());return Math.max(0,t)}function S(e,t,n,r,i){let a=RegExp(`{{\\s*(${n})\\s*(([-+]\\d+)([ymwdhs]))?\\s*(:.+?)?}}`,`gi`),o=window.moment();return e.replace(a,(e,n,a,s,c,l)=>{let u=t.clone();return i&&u.startOf(i),u.set({hour:o.get(`hour`),minute:o.get(`minute`),second:o.get(`second`)}),a&&u.add(parseInt(s,10),c),l?u.format(l.substring(1).trim()):u.format(r)})}function C(e,t,n,r,i){let o=i.replace(/{{\s*date\s*}}/gi,e).replace(/{{\s*time\s*}}/gi,window.moment().format(`HH:mm`)).replace(/{{\s*title\s*}}/gi,e);return t===`day`&&(o=o.replace(/{{\s*yesterday\s*}}/gi,n.clone().subtract(1,`day`).format(r)).replace(/{{\s*tomorrow\s*}}/gi,n.clone().add(1,`d`).format(r)),o=S(o,n,`date|time`,r)),t===`week`&&(o=o.replace(RegExp(`{{\\s*(${a.join(`|`)})\\s*:(.*?)}}`,`gi`),(e,t,r)=>{let i=x(t);return n.weekday(i).format(r.trim())})),(t===`month`||t===`year`)&&(o=S(o,n,t,r,t)),o}async function ee(t,n,r){if(!n||n===`/`)return``;let{metadataCache:i,vault:a}=t,o=(0,e.normalizePath)(n);try{let e=i.getFirstLinkpathDest(o,``);return e?a.cachedRead(e):``}catch(t){return console.error(`[Periodic Notes] Failed to read the ${r} note template '${o}'`,t),new e.Notice(`Failed to read the ${r} note template`),``}}async function te(e,t,n,r){let i=l(n,r.granularity),a=await ee(e,n.granularities[r.granularity].templatePath,r.granularity),o=C(t.basename,r.granularity,r.date,i,a);await e.vault.modify(t,o)}async function ne(t,n,r){let i=(0,e.normalizePath)(y(r.folder??``,n.endsWith(`.md`)?n:`${n}.md`));return await re(t,i),i}async function re(e,t){let n=t.replace(/\\/g,`/`).split(`/`);n.pop();let r=``;for(let t of n)r=r?`${r}/${t}`:t,e.vault.getAbstractFileByPath(r)||await e.vault.createFolder(r)}function ie(e,t){return`${e}:${t.clone().startOf(e).toISOString()}`}function ae(e){let t=e.extension.length+1;return e.path.slice(0,-t)}function oe(e,t,n){if(v(t,n)===`fragile-basename`){let n=ae(e),r=(p(t).match(/\//g)?.length??0)+1;return n.split(`/`).slice(-r).join(`/`)}return e.basename}var se=class extends e.Component{byPath;byKey;sortedByGranularity;dirtyGranularities;constructor(t,n){super(),this.app=t,this.plugin=n,this.byPath=new Map,this.byKey=new Map,this.sortedByGranularity=new Map,this.dirtyGranularities=new Set(c),this.app.workspace.onLayoutReady(()=>{console.info(`[Periodic Notes] initializing cache`),this.initialize(),this.registerEvent(this.app.vault.on(`create`,t=>{t instanceof e.TFile&&this.resolve(t,`create`)})),this.registerEvent(this.app.vault.on(`delete`,t=>{t instanceof e.TFile&&this.remove(t.path)})),this.registerEvent(this.app.vault.on(`rename`,this.onRename,this)),this.registerEvent(this.app.metadataCache.on(`changed`,this.onMetadataChanged,this)),this.registerEvent(this.app.workspace.on(`periodic-notes:settings-updated`,this.reset,this))})}reset(){console.info(`[Periodic Notes] resetting cache`),this.byPath.clear(),this.byKey.clear(),this.sortedByGranularity.clear(),this.dirtyGranularities=new Set(c),this.initialize()}initialize(){let t=this.plugin.settings,n=new Set,r=(t,i)=>{if(!n.has(t)){n.add(t);for(let n of t.children)n instanceof e.TFile?i(n):n instanceof e.TFolder&&r(n,i)}},i=f(t);for(let n of i){let i=t.granularities[n].folder||`/`,a=this.app.vault.getAbstractFileByPath(i);a instanceof e.TFolder&&r(a,t=>{if(t instanceof e.TFile){this.resolve(t,`initialize`);let e=this.app.metadataCache.getFileCache(t);e&&this.onMetadataChanged(t,``,e)}})}}onMetadataChanged(t,n,r){let i=this.plugin.settings,a=f(i);if(a.length!==0)for(let n of a){let a=i.granularities[n].folder||`/`;if(!t.path.startsWith(a===`/`?``:`${a}/`))continue;let o=(0,e.parseFrontMatterEntry)(r.frontmatter,n);if(!o)continue;let s=l(i,n);if(typeof o==`string`){let e=window.moment(o,s,!0);e.isValid()&&this.set({filePath:t.path,date:e,granularity:n,match:`frontmatter`});return}}}onRename(t,n){t instanceof e.TFile&&(this.remove(n),this.resolve(t,`rename`))}resolve(t,n=`create`){let r=this.plugin.settings,i=f(r);if(i.length===0)return;let a=this.byPath.get(t.path);if(!(a&&a.match===`frontmatter`))for(let a of i){let i=r.granularities[a].folder||`/`;if(!t.path.startsWith(i===`/`?``:`${i}/`))continue;let o=u(r,a),s=oe(t,o[0],a),c=window.moment(s,o,!0);if(c.isValid()){let i={filePath:t.path,date:c,granularity:a,match:`filename`};this.set(i),n===`create`&&t.stat.size===0&&te(this.app,t,r,i).catch(n=>{console.error(`[Periodic Notes] failed to apply template`,n),new e.Notice(`Periodic Notes: failed to apply template to "${t.path}". See console for details.`)}),this.app.workspace.trigger(`periodic-notes:resolve`,a,t);return}}}set(e){let t=ie(e.granularity,e.date),n=this.byPath.get(e.filePath);if(n){let e=ie(n.granularity,n.date);e!==t&&(this.byKey.delete(e),this.dirtyGranularities.add(n.granularity))}let r=this.byKey.get(t);r&&r.filePath!==e.filePath&&this.byPath.delete(r.filePath);let i=!this.byKey.has(t);this.byPath.set(e.filePath,e),this.byKey.set(t,e),i&&this.dirtyGranularities.add(e.granularity)}remove(e){let t=this.byPath.get(e);t&&(this.byKey.delete(ie(t.granularity,t.date)),this.byPath.delete(e),this.dirtyGranularities.add(t.granularity))}getPeriodicNote(t,n){let r=ie(t,n),i=this.byKey.get(r);if(!i)return null;let a=this.app.vault.getAbstractFileByPath(i.filePath);return a instanceof e.TFile?a:(this.remove(i.filePath),null)}isPeriodic(e,t){let n=this.byPath.get(e);return n?t?t===n.granularity:!0:!1}find(e){return e?this.byPath.get(e)??null:null}findAdjacent(e,n){let r=this.find(e);if(!r)return null;let i=t(this.getSortedKeys(r.granularity),ie(r.granularity,r.date),n);return i?this.byKey.get(i)??null:null}getSortedKeys(e){if(!this.dirtyGranularities.has(e)){let t=this.sortedByGranularity.get(e);if(t)return t}let t=`${e}:`,n=[];for(let e of this.byKey.keys())e.startsWith(t)&&n.push(e);return n.sort(),this.sortedByGranularity.set(e,n),this.dirtyGranularities.delete(e),n}},ce=Array.isArray,le=Array.prototype.indexOf,ue=Array.prototype.includes,de=Array.from,fe=Object.defineProperty,pe=Object.getOwnPropertyDescriptor,me=Object.getOwnPropertyDescriptors,he=Object.prototype,ge=Array.prototype,_e=Object.getPrototypeOf,ve=Object.isExtensible;function ye(e){return typeof e==`function`}var be=()=>{};function xe(e){for(var t=0;t<e.length;t++)e[t]()}function Se(){var e,t;return{promise:new Promise((n,r)=>{e=n,t=r}),resolve:e,reject:t}}var w=1024,T=2048,Ce=4096,we=8192,Te=16384,Ee=32768,De=1<<25,Oe=65536,ke=1<<19,Ae=1<<20,je=1<<25,Me=65536,Ne=1<<21,Pe=1<<22,Fe=1<<23,Ie=Symbol(`$state`),Le=Symbol(`legacy props`),Re=Symbol(``),ze=new class extends Error{name=`StaleReactionError`;message="The reaction that called `getAbortSignal()` was re-run or destroyed"},Be=!!globalThis.document?.contentType&&globalThis.document.contentType.includes(`xml`);function Ve(e){throw Error(`https://svelte.dev/e/lifecycle_outside_component`)}function He(){throw Error(`https://svelte.dev/e/async_derived_orphan`)}function Ue(e,t,n){throw Error(`https://svelte.dev/e/each_key_duplicate`)}function We(e){throw Error(`https://svelte.dev/e/effect_in_teardown`)}function Ge(){throw Error(`https://svelte.dev/e/effect_in_unowned_derived`)}function Ke(e){throw Error(`https://svelte.dev/e/effect_orphan`)}function qe(){throw Error(`https://svelte.dev/e/effect_update_depth_exceeded`)}function Je(e){throw Error(`https://svelte.dev/e/props_invalid_value`)}function Ye(){throw Error(`https://svelte.dev/e/set_context_after_init`)}function Xe(){throw Error(`https://svelte.dev/e/state_descriptors_fixed`)}function Ze(){throw Error(`https://svelte.dev/e/state_prototype_fixed`)}function Qe(){throw Error(`https://svelte.dev/e/state_unsafe_mutation`)}function $e(){throw Error(`https://svelte.dev/e/svelte_boundary_reset_onerror`)}var et={},E=Symbol(),tt=`http://www.w3.org/1999/xhtml`;function nt(e){console.warn(`https://svelte.dev/e/hydration_mismatch`)}function rt(){console.warn(`https://svelte.dev/e/svelte_boundary_reset_noop`)}var D=!1;function it(e){D=e}var O;function k(e){if(e===null)throw nt(),et;return O=e}function at(){return k(Rn(O))}function A(e){if(D){if(Rn(O)!==null)throw nt(),et;O=e}}function ot(e=1){if(D){for(var t=e,n=O;t--;)n=Rn(n);O=n}}function st(e=!0){for(var t=0,n=O;;){if(n.nodeType===8){var r=n.data;if(r===`]`){if(t===0)return n;--t}else (r===`[`||r===`[!`||r[0]===`[`&&!isNaN(Number(r.slice(1))))&&(t+=1)}var i=Rn(n);e&&n.remove(),n=i}}function ct(e){if(!e||e.nodeType!==8)throw nt(),et;return e.data}function lt(e){return e===this.v}function ut(e,t){return e==e?e!==t||typeof e==`object`&&!!e||typeof e==`function`:t==t}function dt(e){return!ut(e,this.v)}var ft=!1,pt=!1,j=null;function mt(e){j=e}function ht(e){return bt(`getContext`).get(e)}function gt(e,t){let n=bt(`setContext`);if(ft){var r=G.f;!H&&r&32&&!j.i||Ye()}return n.set(e,t),t}function _t(e,t=!1,n){j={p:j,i:!1,c:null,e:null,s:e,x:null,r:G,l:pt&&!t?{s:null,u:null,$:[]}:null}}function vt(e){var t=j,n=t.e;if(n!==null){t.e=null;for(var r of n)Zn(r)}return e!==void 0&&(t.x=e),t.i=!0,j=t.p,e??{}}function yt(){return!pt||j!==null&&j.l===null}function bt(e){return j===null&&Ve(e),j.c??=new Map(xt(j)||void 0)}function xt(e){let t=e.p;for(;t!==null;){let e=t.c;if(e!==null)return e;t=t.p}return null}var St=[];function Ct(){var e=St;St=[],xe(e)}function wt(e){if(St.length===0&&!Ht){var t=St;queueMicrotask(()=>{t===St&&Ct()})}St.push(e)}function Tt(e){var t=G;if(t===null)return H.f|=Fe,e;if(!(t.f&32768)&&!(t.f&4))throw e;Et(e,t)}function Et(e,t){for(;t!==null;){if(t.f&128){if(!(t.f&32768))throw e;try{t.b.error(e);return}catch(t){e=t}}t=t.parent}throw e}var Dt=~(T|Ce|w);function M(e,t){e.f=e.f&Dt|t}function Ot(e){e.f&512||e.deps===null?M(e,w):M(e,Ce)}function kt(e){if(e!==null)for(let t of e)!(t.f&2)||!(t.f&65536)||(t.f^=Me,kt(t.deps))}function At(e,t,n){e.f&2048?t.add(e):e.f&4096&&n.add(e),kt(e.deps),M(e,w)}function jt(e,t,n){if(e==null)return t(void 0),n&&n(void 0),be;let r=Pr(()=>e.subscribe(t,n));return r.unsubscribe?()=>r.unsubscribe():r}var Mt=[];function Nt(e,t=be){let n=null,r=new Set;function i(t){if(ut(e,t)&&(e=t,n)){let t=!Mt.length;for(let t of r)t[1](),Mt.push(t,e);if(t){for(let e=0;e<Mt.length;e+=2)Mt[e][0](Mt[e+1]);Mt.length=0}}}function a(t){i(t(e))}function o(o,s=be){let c=[o,s];return r.add(c),r.size===1&&(n=t(i,a)||be),o(e),()=>{r.delete(c),r.size===0&&n&&(n(),n=null)}}return{set:i,update:a,subscribe:o}}function Pt(e){let t;return jt(e,e=>t=e)(),t}var Ft=!1,It=!1,Lt=Symbol();function Rt(e,t,n){let r=n[t]??={store:null,source:Tn(void 0),unsubscribe:be};if(r.store!==e&&!(Lt in n))if(r.unsubscribe(),r.store=e??null,e==null)r.source.v=void 0,r.unsubscribe=be;else{var i=!0;r.unsubscribe=jt(e,e=>{i?r.source.v=e:R(r.source,e)}),i=!1}return e&&Lt in n?Pt(e):X(r.source)}function zt(){let e={};function t(){Yn(()=>{for(var t in e)e[t].unsubscribe();fe(e,Lt,{enumerable:!1,value:!0})})}return[e,t]}function Bt(e){var t=It;try{return It=!1,[e(),It]}finally{It=t}}var N=new Set,P=null,F=null,Vt=null,Ht=!1,Ut=!1,Wt=null,Gt=null,Kt=0,qt=1,Jt=class e{id=qt++;current=new Map;previous=new Map;#e=new Set;#t=new Set;#n=new Map;#r=new Map;#i=null;#a=[];#o=[];#s=new Set;#c=new Set;#l=new Map;is_fork=!1;#u=!1;#d=new Set;#f(){return this.is_fork||this.#r.size>0}#p(){for(let n of this.#d)for(let r of n.#r.keys()){for(var e=!1,t=r;t.parent!==null;){if(this.#l.has(t)){e=!0;break}t=t.parent}if(!e)return!0}return!1}skip_effect(e){this.#l.has(e)||this.#l.set(e,{d:[],m:[]})}unskip_effect(e){var t=this.#l.get(e);if(t){this.#l.delete(e);for(var n of t.d)M(n,T),this.schedule(n);for(n of t.m)M(n,Ce),this.schedule(n)}}#m(){if(Kt++>1e3&&(N.delete(this),Yt()),!this.#f()){for(let e of this.#s)this.#c.delete(e),M(e,T),this.schedule(e);for(let e of this.#c)M(e,Ce),this.schedule(e)}let t=this.#a;this.#a=[],this.apply();var n=Wt=[],r=[],i=Gt=[];for(let e of t)try{this.#h(e,n,r)}catch(t){throw nn(e),t}if(P=null,i.length>0){var a=e.ensure();for(let e of i)a.schedule(e)}if(Wt=null,Gt=null,this.#f()||this.#p()){this.#g(r),this.#g(n);for(let[e,t]of this.#l)tn(e,t)}else{this.#n.size===0&&N.delete(this),this.#s.clear(),this.#c.clear();for(let e of this.#e)e(this);this.#e.clear(),Zt(r),Zt(n),this.#i?.resolve()}var o=P;if(this.#a.length>0){let e=o??=this;e.#a.push(...this.#a.filter(t=>!e.#a.includes(t)))}o!==null&&(N.add(o),o.#m()),N.has(this)||this.#_()}#h(e,t,n){e.f^=w;for(var r=e.first;r!==null;){var i=r.f,a=(i&96)!=0;if(!(a&&i&1024||i&8192||this.#l.has(r))&&r.fn!==null){a?r.f^=w:i&4?t.push(r):ft&&i&16777224?n.push(r):Er(r)&&(i&16&&this.#c.add(r),jr(r));var o=r.first;if(o!==null){r=o;continue}}for(;r!==null;){var s=r.next;if(s!==null){r=s;break}r=r.parent}}}#g(e){for(var t=0;t<e.length;t+=1)At(e[t],this.#s,this.#c)}capture(e,t,n=!1){t!==E&&!this.previous.has(e)&&this.previous.set(e,t),e.f&8388608||(this.current.set(e,[e.v,n]),F?.set(e,e.v))}activate(){P=this}deactivate(){P=null,F=null}flush(){try{Ut=!0,P=this,this.#m()}finally{Kt=0,Vt=null,Wt=null,Gt=null,Ut=!1,P=null,F=null,Sn.clear()}}discard(){for(let e of this.#t)e(this);this.#t.clear(),N.delete(this)}register_created_effect(e){this.#o.push(e)}#_(){for(let l of N){var e=l.id<this.id,t=[];for(let[r,[i,a]]of this.current){if(l.current.has(r)){var n=l.current.get(r)[0];if(e&&i!==n)l.current.set(r,[i,a]);else continue}t.push(r)}var r=[...l.current.keys()].filter(e=>!this.current.has(e));if(r.length===0)e&&l.discard();else if(t.length>0){l.activate();var i=new Set,a=new Map;for(var o of t)Qt(o,r,i,a);a=new Map;var s=[...l.current.keys()].filter(e=>this.current.has(e)?this.current.get(e)[0]!==e:!0);for(let e of this.#o)!(e.f&155648)&&$t(e,s,a)&&(e.f&4194320?(M(e,T),l.schedule(e)):l.#s.add(e));if(l.#a.length>0){l.apply();for(var c of l.#a)l.#h(c,[],[]);l.#a=[]}l.deactivate()}}for(let e of N)e.#d.has(this)&&(e.#d.delete(this),e.#d.size===0&&!e.#f()&&(e.activate(),e.#m()))}increment(e,t){let n=this.#n.get(t)??0;if(this.#n.set(t,n+1),e){let e=this.#r.get(t)??0;this.#r.set(t,e+1)}}decrement(e,t,n){let r=this.#n.get(t)??0;if(r===1?this.#n.delete(t):this.#n.set(t,r-1),e){let e=this.#r.get(t)??0;e===1?this.#r.delete(t):this.#r.set(t,e-1)}this.#u||n||(this.#u=!0,wt(()=>{this.#u=!1,this.flush()}))}transfer_effects(e,t){for(let t of e)this.#s.add(t);for(let e of t)this.#c.add(e);e.clear(),t.clear()}oncommit(e){this.#e.add(e)}ondiscard(e){this.#t.add(e)}settled(){return(this.#i??=Se()).promise}static ensure(){if(P===null){let t=P=new e;Ut||(N.add(P),Ht||wt(()=>{P===t&&t.flush()}))}return P}apply(){if(!ft||!this.is_fork&&N.size===1){F=null;return}F=new Map;for(let[e,[t]]of this.current)F.set(e,t);for(let n of N)if(!(n===this||n.is_fork)){var e=!1,t=!1;if(n.id<this.id)for(let[r,[,i]]of n.current)i||(e||=this.current.has(r),t||=!this.current.has(r));if(e&&t)this.#d.add(n);else for(let[e,t]of n.previous)F.has(e)||F.set(e,t)}}schedule(e){if(Vt=e,e.b?.is_pending&&e.f&16777228&&!(e.f&32768)){e.b.defer_effect(e);return}for(var t=e;t.parent!==null;){t=t.parent;var n=t.f;if(Wt!==null&&t===G&&(ft||(H===null||!(H.f&2))&&!Ft))return;if(n&96){if(!(n&1024))return;t.f^=w}}this.#a.push(t)}};function Yt(){try{qe()}catch(e){Et(e,Vt)}}var Xt=null;function Zt(e){var t=e.length;if(t!==0){for(var n=0;n<t;){var r=e[n++];if(!(r.f&24576)&&Er(r)&&(Xt=new Set,jr(r),r.deps===null&&r.first===null&&r.nodes===null&&r.teardown===null&&r.ac===null&&cr(r),Xt?.size>0)){Sn.clear();for(let e of Xt){if(e.f&24576)continue;let t=[e],n=e.parent;for(;n!==null;)Xt.has(n)&&(Xt.delete(n),t.push(n)),n=n.parent;for(let e=t.length-1;e>=0;e--){let n=t[e];n.f&24576||jr(n)}}Xt.clear()}}Xt=null}}function Qt(e,t,n,r){if(!n.has(e)&&(n.add(e),e.reactions!==null))for(let i of e.reactions){let e=i.f;e&2?Qt(i,t,n,r):e&4194320&&!(e&2048)&&$t(i,t,r)&&(M(i,T),en(i))}}function $t(e,t,n){let r=n.get(e);if(r!==void 0)return r;if(e.deps!==null)for(let r of e.deps){if(ue.call(t,r))return!0;if(r.f&2&&$t(r,t,n))return n.set(r,!0),!0}return n.set(e,!1),!1}function en(e){P.schedule(e)}function tn(e,t){if(!(e.f&32&&e.f&1024)){e.f&2048?t.d.push(e):e.f&4096&&t.m.push(e),M(e,w);for(var n=e.first;n!==null;)tn(n,t),n=n.next}}function nn(e){M(e,w);for(var t=e.first;t!==null;)nn(t),t=t.next}function rn(e){let t=0,n=wn(0),r;return()=>{Jn()&&(X(n),tr(()=>(t===0&&(r=Pr(()=>e(()=>On(n)))),t+=1,()=>{wt(()=>{--t,t===0&&(r?.(),r=void 0,On(n))})})))}}var an=Oe|ke;function on(e,t,n,r){new sn(e,t,n,r)}var sn=class{parent;is_pending=!1;transform_error;#e;#t=D?O:null;#n;#r;#i;#a=null;#o=null;#s=null;#c=null;#l=0;#u=0;#d=!1;#f=new Set;#p=new Set;#m=null;#h=rn(()=>(this.#m=wn(this.#l),()=>{this.#m=null}));constructor(e,t,n,r){this.#e=e,this.#n=t,this.#r=e=>{var t=G;t.b=this,t.f|=128,n(e)},this.parent=G.b,this.transform_error=r??this.parent?.transform_error??(e=>e),this.#i=rr(()=>{if(D){let e=this.#t;at();let t=e.data===`[!`;if(e.data.startsWith(`[?`)){let t=JSON.parse(e.data.slice(2));this.#_(t)}else t?this.#v():this.#g()}else this.#y()},an),D&&(this.#e=O)}#g(){try{this.#a=B(()=>this.#r(this.#e))}catch(e){this.error(e)}}#_(e){let t=this.#n.failed;t&&(this.#s=B(()=>{t(this.#e,()=>e,()=>()=>{})}))}#v(){let e=this.#n.pending;e&&(this.is_pending=!0,this.#o=B(()=>e(this.#e)),wt(()=>{var e=this.#c=document.createDocumentFragment(),t=In();e.append(t),this.#a=this.#x(()=>B(()=>this.#r(t))),this.#u===0&&(this.#e.before(e),this.#c=null,lr(this.#o,()=>{this.#o=null}),this.#b(P))}))}#y(){try{if(this.is_pending=this.has_pending_snippet(),this.#u=0,this.#l=0,this.#a=B(()=>{this.#r(this.#e)}),this.#u>0){var e=this.#c=document.createDocumentFragment();pr(this.#a,e);let t=this.#n.pending;this.#o=B(()=>t(this.#e))}else this.#b(P)}catch(e){this.error(e)}}#b(e){this.is_pending=!1,e.transfer_effects(this.#f,this.#p)}defer_effect(e){At(e,this.#f,this.#p)}is_rendered(){return!this.is_pending&&(!this.parent||this.parent.is_rendered())}has_pending_snippet(){return!!this.#n.pending}#x(e){var t=G,n=H,r=j;vr(this.#i),W(this.#i),mt(this.#i.ctx);try{return Jt.ensure(),e()}catch(e){return Tt(e),null}finally{vr(t),W(n),mt(r)}}#S(e,t){if(!this.has_pending_snippet()){this.parent&&this.parent.#S(e,t);return}this.#u+=e,this.#u===0&&(this.#b(t),this.#o&&lr(this.#o,()=>{this.#o=null}),this.#c&&=(this.#e.before(this.#c),null))}update_pending_count(e,t){this.#S(e,t),this.#l+=e,!(!this.#m||this.#d)&&(this.#d=!0,wt(()=>{this.#d=!1,this.#m&&En(this.#m,this.#l)}))}get_effect_pending(){return this.#h(),X(this.#m)}error(e){var t=this.#n.onerror;let n=this.#n.failed;if(!t&&!n)throw e;this.#a&&=(V(this.#a),null),this.#o&&=(V(this.#o),null),this.#s&&=(V(this.#s),null),D&&(k(this.#t),ot(),k(st()));var r=!1,i=!1;let a=()=>{if(r){rt();return}r=!0,i&&$e(),this.#s!==null&&lr(this.#s,()=>{this.#s=null}),this.#x(()=>{this.#y()})},o=e=>{try{i=!0,t?.(e,a),i=!1}catch(e){Et(e,this.#i&&this.#i.parent)}n&&(this.#s=this.#x(()=>{try{return B(()=>{var t=G;t.b=this,t.f|=128,n(this.#e,()=>e,()=>a)})}catch(e){return Et(e,this.#i.parent),null}}))};wt(()=>{var t;try{t=this.transform_error(e)}catch(e){Et(e,this.#i&&this.#i.parent);return}typeof t==`object`&&t&&typeof t.then==`function`?t.then(o,e=>Et(e,this.#i&&this.#i.parent)):o(t)})}};function cn(e,t,n,r){let i=yt()?fn:mn;var a=e.filter(e=>!e.settled);if(n.length===0&&a.length===0){r(t.map(i));return}var o=G,s=ln(),c=a.length===1?a[0].promise:a.length>1?Promise.all(a.map(e=>e.promise)):null;function l(e){s();try{r(e)}catch(e){o.f&16384||Et(e,o)}un()}if(n.length===0){c.then(()=>l(t.map(i)));return}var u=dn();function d(){Promise.all(n.map(e=>pn(e))).then(e=>l([...t.map(i),...e])).catch(e=>Et(e,o)).finally(()=>u())}c?c.then(()=>{s(),d(),un()}):d()}function ln(){var e=G,t=H,n=j,r=P;return function(i=!0){vr(e),W(t),mt(n),i&&!(e.f&16384)&&(r?.activate(),r?.apply())}}function un(e=!0){vr(null),W(null),mt(null),e&&P?.deactivate()}function dn(){var e=G,t=e.b,n=P,r=t.is_rendered();return t.update_pending_count(1,n),n.increment(r,e),(i=!1)=>{t.update_pending_count(-1,n),n.decrement(r,e,i)}}function fn(e){var t=2|T,n=H!==null&&H.f&2?H:null;return G!==null&&(G.f|=ke),{ctx:j,deps:null,effects:null,equals:lt,f:t,fn:e,reactions:null,rv:0,v:E,wv:0,parent:n??G,ac:null}}function pn(e,t,n){let r=G;r===null&&He();var i=void 0,a=wn(E),o=!H,s=new Map;return er(()=>{var t=G,n=Se();i=n.promise;try{Promise.resolve(e()).then(n.resolve,n.reject).finally(un)}catch(e){n.reject(e),un()}var c=P;if(o){if(t.f&32768)var l=dn();if(r.b.is_rendered())s.get(c)?.reject(ze),s.delete(c);else{for(let e of s.values())e.reject(ze);s.clear()}s.set(c,n)}let u=(e,n=void 0)=>{if(l&&l(n===ze),!(n===ze||t.f&16384)){if(c.activate(),n)a.f|=Fe,En(a,n);else{a.f&8388608&&(a.f^=Fe),En(a,e);for(let[e,t]of s){if(s.delete(e),e===c)break;t.reject(ze)}}c.deactivate()}};n.promise.then(u,e=>u(null,e||`unknown`))}),Yn(()=>{for(let e of s.values())e.reject(ze)}),new Promise(e=>{function t(n){function r(){n===i?e(a):t(i)}n.then(r,r)}t(i)})}function I(e){let t=fn(e);return ft||yr(t),t}function mn(e){let t=fn(e);return t.equals=dt,t}function hn(e){var t=e.effects;if(t!==null){e.effects=null;for(var n=0;n<t.length;n+=1)V(t[n])}}function gn(e){for(var t=e.parent;t!==null;){if(!(t.f&2))return t.f&16384?null:t;t=t.parent}return null}function _n(e){var t,n=G;vr(gn(e));try{e.f&=~Me,hn(e),t=Or(e)}finally{vr(n)}return t}function vn(e){var t=e.v,n=_n(e);if(!e.equals(n)&&(e.wv=Tr(),(!P?.is_fork||e.deps===null)&&(e.v=n,P?.capture(e,t,!0),e.deps===null))){M(e,w);return}gr||(F===null?Ot(e):(Jn()||P?.is_fork)&&F.set(e,n))}function yn(e){if(e.effects!==null)for(let t of e.effects)(t.teardown||t.ac)&&(t.teardown?.(),t.ac?.abort(ze),t.teardown=be,t.ac=null,Ar(t,0),ar(t))}function bn(e){if(e.effects!==null)for(let t of e.effects)t.teardown&&jr(t)}var xn=new Set,Sn=new Map,Cn=!1;function wn(e,t){return{f:0,v:e,reactions:null,equals:lt,rv:0,wv:0}}function L(e,t){let n=wn(e,t);return yr(n),n}function Tn(e,t=!1,n=!0){let r=wn(e);return t||(r.equals=dt),pt&&n&&j!==null&&j.l!==null&&(j.l.s??=[]).push(r),r}function R(e,t,n=!1){return H!==null&&(!U||H.f&131072)&&yt()&&H.f&4325394&&(K===null||!ue.call(K,e))&&Qe(),En(e,n?An(t):t,Gt)}function En(e,t,n=null){if(!e.equals(t)){var r=e.v;gr?Sn.set(e,t):Sn.set(e,r),e.v=t;var i=Jt.ensure();if(i.capture(e,r),e.f&2){let t=e;e.f&2048&&_n(t),F===null&&Ot(t)}e.wv=Tr(),kn(e,T,n),yt()&&G!==null&&G.f&1024&&!(G.f&96)&&(Y===null?br([e]):Y.push(e)),!i.is_fork&&xn.size>0&&!Cn&&Dn()}return t}function Dn(){Cn=!1;for(let e of xn)e.f&1024&&M(e,Ce),Er(e)&&jr(e);xn.clear()}function On(e){R(e,e.v+1)}function kn(e,t,n){var r=e.reactions;if(r!==null)for(var i=yt(),a=r.length,o=0;o<a;o++){var s=r[o],c=s.f;if(!(!i&&s===G)){var l=(c&T)===0;if(l&&M(s,t),c&2){var u=s;F?.delete(u),c&65536||(c&512&&(s.f|=Me),kn(u,Ce,n))}else if(l){var d=s;c&16&&Xt!==null&&Xt.add(d),n===null?en(d):n.push(d)}}}}function An(e){if(typeof e!=`object`||!e||Ie in e)return e;let t=_e(e);if(t!==he&&t!==ge)return e;var n=new Map,r=ce(e),i=L(0),a=null,o=Cr,s=e=>{if(Cr===o)return e();var t=H,n=Cr;W(null),wr(o);var r=e();return W(t),wr(n),r};return r&&n.set(`length`,L(e.length,a)),new Proxy(e,{defineProperty(e,t,r){(!(`value`in r)||r.configurable===!1||r.enumerable===!1||r.writable===!1)&&Xe();var i=n.get(t);return i===void 0?s(()=>{var e=L(r.value,a);return n.set(t,e),e}):R(i,r.value,!0),!0},deleteProperty(e,t){var r=n.get(t);if(r===void 0){if(t in e){let e=s(()=>L(E,a));n.set(t,e),On(i)}}else R(r,E),On(i);return!0},get(t,r,i){if(r===Ie)return e;var o=n.get(r),c=r in t;if(o===void 0&&(!c||pe(t,r)?.writable)&&(o=s(()=>L(An(c?t[r]:E),a)),n.set(r,o)),o!==void 0){var l=X(o);return l===E?void 0:l}return Reflect.get(t,r,i)},getOwnPropertyDescriptor(e,t){var r=Reflect.getOwnPropertyDescriptor(e,t);if(r&&`value`in r){var i=n.get(t);i&&(r.value=X(i))}else if(r===void 0){var a=n.get(t),o=a?.v;if(a!==void 0&&o!==E)return{enumerable:!0,configurable:!0,value:o,writable:!0}}return r},has(e,t){if(t===Ie)return!0;var r=n.get(t),i=r!==void 0&&r.v!==E||Reflect.has(e,t);return(r!==void 0||G!==null&&(!i||pe(e,t)?.writable))&&(r===void 0&&(r=s(()=>L(i?An(e[t]):E,a)),n.set(t,r)),X(r)===E)?!1:i},set(e,t,o,c){var l=n.get(t),u=t in e;if(r&&t===`length`)for(var d=o;d<l.v;d+=1){var f=n.get(d+``);f===void 0?d in e&&(f=s(()=>L(E,a)),n.set(d+``,f)):R(f,E)}if(l===void 0)(!u||pe(e,t)?.writable)&&(l=s(()=>L(void 0,a)),R(l,An(o)),n.set(t,l));else{u=l.v!==E;var p=s(()=>An(o));R(l,p)}var m=Reflect.getOwnPropertyDescriptor(e,t);if(m?.set&&m.set.call(c,o),!u){if(r&&typeof t==`string`){var h=n.get(`length`),g=Number(t);Number.isInteger(g)&&g>=h.v&&R(h,g+1)}On(i)}return!0},ownKeys(e){X(i);var t=Reflect.ownKeys(e).filter(e=>{var t=n.get(e);return t===void 0||t.v!==E});for(var[r,a]of n)a.v!==E&&!(r in e)&&t.push(r);return t},setPrototypeOf(){Ze()}})}var jn,Mn,Nn,Pn;function Fn(){if(jn===void 0){jn=window,Mn=/Firefox/.test(navigator.userAgent);var e=Element.prototype,t=Node.prototype,n=Text.prototype;Nn=pe(t,`firstChild`).get,Pn=pe(t,`nextSibling`).get,ve(e)&&(e.__click=void 0,e.__className=void 0,e.__attributes=null,e.__style=void 0,e.__e=void 0),ve(n)&&(n.__t=void 0)}}function In(e=``){return document.createTextNode(e)}function Ln(e){return Nn.call(e)}function Rn(e){return Pn.call(e)}function z(e,t){if(!D)return Ln(e);var n=Ln(O);if(n===null)n=O.appendChild(In());else if(t&&n.nodeType!==3){var r=In();return n?.before(r),k(r),r}return t&&Un(n),k(n),n}function zn(e,t=1,n=!1){let r=D?O:e;for(var i;t--;)i=r,r=Rn(r);if(!D)return r;if(n){if(r?.nodeType!==3){var a=In();return r===null?i?.after(a):r.before(a),k(a),a}Un(r)}return k(r),r}function Bn(e){e.textContent=``}function Vn(){return!ft||Xt!==null?!1:(G.f&Ee)!==0}function Hn(e,t,n){let r=n?{is:n}:void 0;return document.createElementNS(t??`http://www.w3.org/1999/xhtml`,e,r)}function Un(e){if(e.nodeValue.length<65536)return;let t=e.nextSibling;for(;t!==null&&t.nodeType===3;)t.remove(),e.nodeValue+=t.nodeValue,t=e.nextSibling}function Wn(e){var t=H,n=G;W(null),vr(null);try{return e()}finally{W(t),vr(n)}}function Gn(e){G===null&&(H===null&&Ke(e),Ge()),gr&&We(e)}function Kn(e,t){var n=t.last;n===null?t.last=t.first=e:(n.next=e,e.prev=n,t.last=e)}function qn(e,t){var n=G;n!==null&&n.f&8192&&(e|=we);var r={ctx:j,deps:null,nodes:null,f:e|T|512,first:null,fn:t,last:null,next:null,parent:n,b:n&&n.b,prev:null,teardown:null,wv:0,ac:null};P?.register_created_effect(r);var i=r;if(e&4)Wt===null?Jt.ensure().schedule(r):Wt.push(r);else if(t!==null){try{jr(r)}catch(e){throw V(r),e}i.deps===null&&i.teardown===null&&i.nodes===null&&i.first===i.last&&!(i.f&524288)&&(i=i.first,e&16&&e&65536&&i!==null&&(i.f|=Oe))}if(i!==null&&(i.parent=n,n!==null&&Kn(i,n),H!==null&&H.f&2&&!(e&64))){var a=H;(a.effects??=[]).push(i)}return r}function Jn(){return H!==null&&!U}function Yn(e){let t=qn(8,null);return M(t,w),t.teardown=e,t}function Xn(e){Gn(`$effect`);var t=G.f;if(!H&&t&32&&!(t&32768)){var n=j;(n.e??=[]).push(e)}else return Zn(e)}function Zn(e){return qn(4|Ae,e)}function Qn(e){Jt.ensure();let t=qn(64|ke,e);return(e={})=>new Promise(n=>{e.outro?lr(t,()=>{V(t),n(void 0)}):(V(t),n(void 0))})}function $n(e){return qn(4,e)}function er(e){return qn(Pe|ke,e)}function tr(e,t=0){return qn(8|t,e)}function nr(e,t=[],n=[],r=[]){cn(r,t,n,t=>{qn(8,()=>e(...t.map(X)))})}function rr(e,t=0){return qn(16|t,e)}function B(e){return qn(32|ke,e)}function ir(e){var t=e.teardown;if(t!==null){let e=gr,n=H;_r(!0),W(null);try{t.call(null)}finally{_r(e),W(n)}}}function ar(e,t=!1){var n=e.first;for(e.first=e.last=null;n!==null;){let e=n.ac;e!==null&&Wn(()=>{e.abort(ze)});var r=n.next;n.f&64?n.parent=null:V(n,t),n=r}}function or(e){for(var t=e.first;t!==null;){var n=t.next;t.f&32||V(t),t=n}}function V(e,t=!0){var n=!1;(t||e.f&262144)&&e.nodes!==null&&e.nodes.end!==null&&(sr(e.nodes.start,e.nodes.end),n=!0),M(e,De),ar(e,t&&!n),Ar(e,0);var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)e.stop();ir(e),e.f^=De,e.f|=Te;var i=e.parent;i!==null&&i.first!==null&&cr(e),e.next=e.prev=e.teardown=e.ctx=e.deps=e.fn=e.nodes=e.ac=e.b=null}function sr(e,t){for(;e!==null;){var n=e===t?null:Rn(e);e.remove(),e=n}}function cr(e){var t=e.parent,n=e.prev,r=e.next;n!==null&&(n.next=r),r!==null&&(r.prev=n),t!==null&&(t.first===e&&(t.first=r),t.last===e&&(t.last=n))}function lr(e,t,n=!0){var r=[];ur(e,r,!0);var i=()=>{n&&V(e),t&&t()},a=r.length;if(a>0){var o=()=>--a||i();for(var s of r)s.out(o)}else i()}function ur(e,t,n){if(!(e.f&8192)){e.f^=we;var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)(e.is_global||n)&&t.push(e);for(var i=e.first;i!==null;){var a=i.next,o=(i.f&65536)!=0||(i.f&32)!=0&&(e.f&16)!=0;ur(i,t,o?n:!1),i=a}}}function dr(e){fr(e,!0)}function fr(e,t){if(e.f&8192){e.f^=we,e.f&1024||(M(e,T),Jt.ensure().schedule(e));for(var n=e.first;n!==null;){var r=n.next,i=(n.f&65536)!=0||(n.f&32)!=0;fr(n,i?t:!1),n=r}var a=e.nodes&&e.nodes.t;if(a!==null)for(let e of a)(e.is_global||t)&&e.in()}}function pr(e,t){if(e.nodes)for(var n=e.nodes.start,r=e.nodes.end;n!==null;){var i=n===r?null:Rn(n);t.append(n),n=i}}var mr=null,hr=!1,gr=!1;function _r(e){gr=e}var H=null,U=!1;function W(e){H=e}var G=null;function vr(e){G=e}var K=null;function yr(e){H!==null&&(!ft||H.f&2)&&(K===null?K=[e]:K.push(e))}var q=null,J=0,Y=null;function br(e){Y=e}var xr=1,Sr=0,Cr=Sr;function wr(e){Cr=e}function Tr(){return++xr}function Er(e){var t=e.f;if(t&2048)return!0;if(t&2&&(e.f&=~Me),t&4096){for(var n=e.deps,r=n.length,i=0;i<r;i++){var a=n[i];if(Er(a)&&vn(a),a.wv>e.wv)return!0}t&512&&F===null&&M(e,w)}return!1}function Dr(e,t,n=!0){var r=e.reactions;if(r!==null&&!(!ft&&K!==null&&ue.call(K,e)))for(var i=0;i<r.length;i++){var a=r[i];a.f&2?Dr(a,t,!1):t===a&&(n?M(a,T):a.f&1024&&M(a,Ce),en(a))}}function Or(e){var t=q,n=J,r=Y,i=H,a=K,o=j,s=U,c=Cr,l=e.f;q=null,J=0,Y=null,H=l&96?null:e,K=null,mt(e.ctx),U=!1,Cr=++Sr,e.ac!==null&&(Wn(()=>{e.ac.abort(ze)}),e.ac=null);try{e.f|=Ne;var u=e.fn,d=u();e.f|=Ee;var f=e.deps,p=P?.is_fork;if(q!==null){var m;if(p||Ar(e,J),f!==null&&J>0)for(f.length=J+q.length,m=0;m<q.length;m++)f[J+m]=q[m];else e.deps=f=q;if(Jn()&&e.f&512)for(m=J;m<f.length;m++)(f[m].reactions??=[]).push(e)}else !p&&f!==null&&J<f.length&&(Ar(e,J),f.length=J);if(yt()&&Y!==null&&!U&&f!==null&&!(e.f&6146))for(m=0;m<Y.length;m++)Dr(Y[m],e);if(i!==null&&i!==e){if(Sr++,i.deps!==null)for(let e=0;e<n;e+=1)i.deps[e].rv=Sr;if(t!==null)for(let e of t)e.rv=Sr;Y!==null&&(r===null?r=Y:r.push(...Y))}return e.f&8388608&&(e.f^=Fe),d}catch(e){return Tt(e)}finally{e.f^=Ne,q=t,J=n,Y=r,H=i,K=a,mt(o),U=s,Cr=c}}function kr(e,t){let n=t.reactions;if(n!==null){var r=le.call(n,e);if(r!==-1){var i=n.length-1;i===0?n=t.reactions=null:(n[r]=n[i],n.pop())}}if(n===null&&t.f&2&&(q===null||!ue.call(q,t))){var a=t;a.f&512&&(a.f^=512,a.f&=~Me),Ot(a),yn(a),Ar(a,0)}}function Ar(e,t){var n=e.deps;if(n!==null)for(var r=t;r<n.length;r++)kr(e,n[r])}function jr(e){var t=e.f;if(!(t&16384)){M(e,w);var n=G,r=hr;G=e,hr=!0;try{t&16777232?or(e):ar(e),ir(e);var i=Or(e);e.teardown=typeof i==`function`?i:null,e.wv=xr}finally{hr=r,G=n}}}function X(e){var t=(e.f&2)!=0;if(mr?.add(e),H!==null&&!U&&!(G!==null&&G.f&16384)&&(K===null||!ue.call(K,e))){var n=H.deps;if(H.f&2097152)e.rv<Sr&&(e.rv=Sr,q===null&&n!==null&&n[J]===e?J++:q===null?q=[e]:q.push(e));else{(H.deps??=[]).push(e);var r=e.reactions;r===null?e.reactions=[H]:ue.call(r,H)||r.push(H)}}if(gr&&Sn.has(e))return Sn.get(e);if(t){var i=e;if(gr){var a=i.v;return(!(i.f&1024)&&i.reactions!==null||Nr(i))&&(a=_n(i)),Sn.set(i,a),a}var o=(i.f&512)==0&&!U&&H!==null&&(hr||(H.f&512)!=0),s=(i.f&Ee)===0;Er(i)&&(o&&(i.f|=512),vn(i)),o&&!s&&(bn(i),Mr(i))}if(F?.has(e))return F.get(e);if(e.f&8388608)throw e.v;return e.v}function Mr(e){if(e.f|=512,e.deps!==null)for(let t of e.deps)(t.reactions??=[]).push(e),t.f&2&&!(t.f&512)&&(bn(t),Mr(t))}function Nr(e){if(e.v===E)return!0;if(e.deps===null)return!1;for(let t of e.deps)if(Sn.has(t)||t.f&2&&Nr(t))return!0;return!1}function Pr(e){var t=U;try{return U=!0,e()}finally{U=t}}[...`allowfullscreen.async.autofocus.autoplay.checked.controls.default.disabled.formnovalidate.indeterminate.inert.ismap.loop.multiple.muted.nomodule.novalidate.open.playsinline.readonly.required.reversed.seamless.selected.webkitdirectory.defer.disablepictureinpicture.disableremoteplayback`.split(`.`)];var Fr=[`touchstart`,`touchmove`];function Ir(e){return Fr.includes(e)}var Lr=Symbol(`events`),Rr=new Set,zr=new Set;function Br(e,t,n,r={}){function i(e){if(r.capture||Wr.call(t,e),!e.cancelBubble)return Wn(()=>n?.call(this,e))}return e.startsWith(`pointer`)||e.startsWith(`touch`)||e===`wheel`?wt(()=>{t.addEventListener(e,i,r)}):t.addEventListener(e,i,r),i}function Vr(e,t,n,r,i){var a={capture:r,passive:i},o=Br(e,t,n,a);(t===document.body||t===window||t===document||t instanceof HTMLMediaElement)&&Yn(()=>{t.removeEventListener(e,o,a)})}function Z(e,t,n){(t[Lr]??={})[e]=n}function Hr(e){for(var t=0;t<e.length;t++)Rr.add(e[t]);for(var n of zr)n(e)}var Ur=null;function Wr(e){var t=this,n=t.ownerDocument,r=e.type,i=e.composedPath?.()||[],a=i[0]||e.target;Ur=e;var o=0,s=Ur===e&&e[Lr];if(s){var c=i.indexOf(s);if(c!==-1&&(t===document||t===window)){e[Lr]=t;return}var l=i.indexOf(t);if(l===-1)return;c<=l&&(o=c)}if(a=i[o]||e.target,a!==t){fe(e,`currentTarget`,{configurable:!0,get(){return a||n}});var u=H,d=G;W(null),vr(null);try{for(var f,p=[];a!==null;){var m=a.assignedSlot||a.parentNode||a.host||null;try{var h=a[Lr]?.[r];h!=null&&(!a.disabled||e.target===a)&&h.call(a,e)}catch(e){f?p.push(e):f=e}if(e.cancelBubble||m===t||m===null)break;a=m}if(f){for(let e of p)queueMicrotask(()=>{throw e});throw f}}finally{e[Lr]=t,delete e.currentTarget,W(u),vr(d)}}}var Gr=globalThis?.window?.trustedTypes&&globalThis.window.trustedTypes.createPolicy(`svelte-trusted-html`,{createHTML:e=>e});function Kr(e){return Gr?.createHTML(e)??e}function qr(e){var t=Hn(`template`);return t.innerHTML=Kr(e.replaceAll(`<!>`,`<!---->`)),t.content}function Jr(e,t){var n=G;n.nodes===null&&(n.nodes={start:e,end:t,a:null,t:null})}function Q(e,t){var n=(t&1)!=0,r=(t&2)!=0,i,a=!e.startsWith(`<!>`);return()=>{if(D)return Jr(O,null),O;i===void 0&&(i=qr(a?e:`<!>`+e),n||(i=Ln(i)));var t=r||Mn?document.importNode(i,!0):i.cloneNode(!0);if(n){var o=Ln(t),s=t.lastChild;Jr(o,s)}else Jr(t,t);return t}}function $(e,t){if(D){var n=G;(!(n.f&32768)||n.nodes.end===null)&&(n.nodes.end=O),at();return}e!==null&&e.before(t)}function Yr(e,t){var n=t==null?``:typeof t==`object`?`${t}`:t;n!==(e.__t??=e.nodeValue)&&(e.__t=n,e.nodeValue=`${n}`)}function Xr(e,t){return Qr(e,t)}var Zr=new Map;function Qr(e,{target:t,anchor:n,props:r={},events:i,context:a,intro:o=!0,transformError:s}){Fn();var c=void 0,l=Qn(()=>{var o=n??t.appendChild(In());on(o,{pending:()=>{}},t=>{_t({});var n=j;if(a&&(n.c=a),i&&(r.$$events=i),D&&Jr(t,null),c=e(t,r)||{},D&&(G.nodes.end=O,O===null||O.nodeType!==8||O.data!==`]`))throw nt(),et;vt()},s);var l=new Set,u=e=>{for(var n=0;n<e.length;n++){var r=e[n];if(!l.has(r)){l.add(r);var i=Ir(r);for(let e of[t,document]){var a=Zr.get(e);a===void 0&&(a=new Map,Zr.set(e,a));var o=a.get(r);o===void 0?(e.addEventListener(r,Wr,{passive:i}),a.set(r,1)):a.set(r,o+1)}}}};return u(de(Rr)),zr.add(u),()=>{for(var e of l)for(let n of[t,document]){var r=Zr.get(n),i=r.get(e);--i==0?(n.removeEventListener(e,Wr),r.delete(e),r.size===0&&Zr.delete(n)):r.set(e,i)}zr.delete(u),o!==n&&o.parentNode?.removeChild(o)}});return $r.set(c,l),c}var $r=new WeakMap;function ei(e,t){let n=$r.get(e);return n?($r.delete(e),n(t)):Promise.resolve()}var ti=class{anchor;#e=new Map;#t=new Map;#n=new Map;#r=new Set;#i=!0;constructor(e,t=!0){this.anchor=e,this.#i=t}#a=e=>{if(this.#e.has(e)){var t=this.#e.get(e),n=this.#t.get(t);if(n)dr(n),this.#r.delete(t);else{var r=this.#n.get(t);r&&(this.#t.set(t,r.effect),this.#n.delete(t),r.fragment.lastChild.remove(),this.anchor.before(r.fragment),n=r.effect)}for(let[t,n]of this.#e){if(this.#e.delete(t),t===e)break;let r=this.#n.get(n);r&&(V(r.effect),this.#n.delete(n))}for(let[e,r]of this.#t){if(e===t||this.#r.has(e))continue;let i=()=>{if(Array.from(this.#e.values()).includes(e)){var t=document.createDocumentFragment();pr(r,t),t.append(In()),this.#n.set(e,{effect:r,fragment:t})}else V(r);this.#r.delete(e),this.#t.delete(e)};this.#i||!n?(this.#r.add(e),lr(r,i,!1)):i()}}};#o=e=>{this.#e.delete(e);let t=Array.from(this.#e.values());for(let[e,n]of this.#n)t.includes(e)||(V(n.effect),this.#n.delete(e))};ensure(e,t){var n=P,r=Vn();if(t&&!this.#t.has(e)&&!this.#n.has(e))if(r){var i=document.createDocumentFragment(),a=In();i.append(a),this.#n.set(e,{effect:B(()=>t(a)),fragment:i})}else this.#t.set(e,B(()=>t(this.anchor)));if(this.#e.set(n,e),r){for(let[t,r]of this.#t)t===e?n.unskip_effect(r):n.skip_effect(r);for(let[t,r]of this.#n)t===e?n.unskip_effect(r.effect):n.skip_effect(r.effect);n.oncommit(this.#a),n.ondiscard(this.#o)}else D&&(this.anchor=O),this.#a(n)}};function ni(e,t,n=!1){var r;D&&(r=O,at());var i=new ti(e),a=n?Oe:0;function o(e,t){if(D){var n=ct(r);if(e!==parseInt(n.substring(1))){var a=st();k(a),i.anchor=a,it(!1),i.ensure(e,t),it(!0);return}}i.ensure(e,t)}rr(()=>{var e=!1;t((t,n=0)=>{e=!0,o(n,t)}),e||o(-1,null)},a)}function ri(e,t){return t}function ii(e,t,n){for(var r=[],i=t.length,a,o=t.length,s=0;s<i;s++){let n=t[s];lr(n,()=>{if(a){if(a.pending.delete(n),a.done.add(n),a.pending.size===0){var t=e.outrogroups;ai(e,de(a.done)),t.delete(a),t.size===0&&(e.outrogroups=null)}}else --o},!1)}if(o===0){var c=r.length===0&&n!==null;if(c){var l=n,u=l.parentNode;Bn(u),u.append(l),e.items.clear()}ai(e,t,!c)}else a={pending:new Set(t),done:new Set},(e.outrogroups??=new Set).add(a)}function ai(e,t,n=!0){var r;if(e.pending.size>0){r=new Set;for(let t of e.pending.values())for(let n of t)r.add(e.items.get(n).e)}for(var i=0;i<t.length;i++){var a=t[i];r?.has(a)?(a.f|=je,pr(a,document.createDocumentFragment())):V(t[i],n)}}var oi;function si(e,t,n,r,i,a=null){var o=e,s=new Map;if(t&4){var c=e;o=D?k(Ln(c)):c.appendChild(In())}D&&at();var l=null,u=mn(()=>{var e=n();return ce(e)?e:e==null?[]:de(e)}),d,f=new Map,p=!0;function m(e){g.effect.f&16384||(g.pending.delete(e),g.fallback=l,li(g,d,o,t,r),l!==null&&(d.length===0?l.f&33554432?(l.f^=je,di(l,null,o)):dr(l):lr(l,()=>{l=null})))}function h(e){g.pending.delete(e)}var g={effect:rr(()=>{d=X(u);var e=d.length;let c=!1;D&&ct(o)===`[!`!=(e===0)&&(o=st(),k(o),it(!1),c=!0);for(var g=new Set,_=P,v=Vn(),y=0;y<e;y+=1){D&&O.nodeType===8&&O.data===`]`&&(o=O,c=!0,it(!1));var b=d[y],x=r(b,y),S=p?null:s.get(x);S?(S.v&&En(S.v,b),S.i&&En(S.i,y),v&&_.unskip_effect(S.e)):(S=ui(s,p?o:oi??=In(),b,x,y,i,t,n),p||(S.e.f|=je),s.set(x,S)),g.add(x)}if(e===0&&a&&!l&&(p?l=B(()=>a(o)):(l=B(()=>a(oi??=In())),l.f|=je)),e>g.size&&Ue(``,``,``),D&&e>0&&k(st()),!p)if(f.set(_,g),v){for(let[e,t]of s)g.has(e)||_.skip_effect(t.e);_.oncommit(m),_.ondiscard(h)}else m(_);c&&it(!0),X(u)}),flags:t,items:s,pending:f,outrogroups:null,fallback:l};p=!1,D&&(o=O)}function ci(e){for(;e!==null&&!(e.f&32);)e=e.next;return e}function li(e,t,n,r,i){var a=(r&8)!=0,o=t.length,s=e.items,c=ci(e.effect.first),l,u=null,d,f=[],p=[],m,h,g,_;if(a)for(_=0;_<o;_+=1)m=t[_],h=i(m,_),g=s.get(h).e,g.f&33554432||(g.nodes?.a?.measure(),(d??=new Set).add(g));for(_=0;_<o;_+=1){if(m=t[_],h=i(m,_),g=s.get(h).e,e.outrogroups!==null)for(let t of e.outrogroups)t.pending.delete(g),t.done.delete(g);if(g.f&8192&&(dr(g),a&&(g.nodes?.a?.unfix(),(d??=new Set).delete(g))),g.f&33554432)if(g.f^=je,g===c)di(g,null,n);else{var v=u?u.next:c;g===e.effect.last&&(e.effect.last=g.prev),g.prev&&(g.prev.next=g.next),g.next&&(g.next.prev=g.prev),fi(e,u,g),fi(e,g,v),di(g,v,n),u=g,f=[],p=[],c=ci(u.next);continue}if(g!==c){if(l!==void 0&&l.has(g)){if(f.length<p.length){var y=p[0],b;u=y.prev;var x=f[0],S=f[f.length-1];for(b=0;b<f.length;b+=1)di(f[b],y,n);for(b=0;b<p.length;b+=1)l.delete(p[b]);fi(e,x.prev,S.next),fi(e,u,x),fi(e,S,y),c=y,u=S,--_,f=[],p=[]}else l.delete(g),di(g,c,n),fi(e,g.prev,g.next),fi(e,g,u===null?e.effect.first:u.next),fi(e,u,g),u=g;continue}for(f=[],p=[];c!==null&&c!==g;)(l??=new Set).add(c),p.push(c),c=ci(c.next);if(c===null)continue}g.f&33554432||f.push(g),u=g,c=ci(g.next)}if(e.outrogroups!==null){for(let t of e.outrogroups)t.pending.size===0&&(ai(e,de(t.done)),e.outrogroups?.delete(t));e.outrogroups.size===0&&(e.outrogroups=null)}if(c!==null||l!==void 0){var C=[];if(l!==void 0)for(g of l)g.f&8192||C.push(g);for(;c!==null;)!(c.f&8192)&&c!==e.fallback&&C.push(c),c=ci(c.next);var ee=C.length;if(ee>0){var te=r&4&&o===0?n:null;if(a){for(_=0;_<ee;_+=1)C[_].nodes?.a?.measure();for(_=0;_<ee;_+=1)C[_].nodes?.a?.fix()}ii(e,C,te)}}a&&wt(()=>{if(d!==void 0)for(g of d)g.nodes?.a?.apply()})}function ui(e,t,n,r,i,a,o,s){var c=o&1?o&16?wn(n):Tn(n,!1,!1):null,l=o&2?wn(i):null;return{v:c,i:l,e:B(()=>(a(t,c??n,l??i,s),()=>{e.delete(r)}))}}function di(e,t,n){if(e.nodes)for(var r=e.nodes.start,i=e.nodes.end,a=t&&!(t.f&33554432)?t.nodes.start:n;r!==null;){var o=Rn(r);if(a.before(r),r===i)return;r=o}}function fi(e,t,n){t===null?e.effect.first=n:t.next=n,n===null?e.effect.last=t:n.prev=t}function pi(e,t){$n(()=>{var n=e.getRootNode(),r=n.host?n:n.head??n.ownerDocument.head;if(!r.querySelector(`#`+t.hash)){let e=Hn(`style`);e.id=t.hash,e.textContent=t.code,r.appendChild(e)}})}var mi=[...` 	
-\r\f\xA0\v﻿`];function hi(e,t,n){var r=e==null?``:``+e;if(t&&(r=r?r+` `+t:t),n){for(var i of Object.keys(n))if(n[i])r=r?r+` `+i:i;else if(r.length)for(var a=i.length,o=0;(o=r.indexOf(i,o))>=0;){var s=o+a;(o===0||mi.includes(r[o-1]))&&(s===r.length||mi.includes(r[s]))?r=(o===0?``:r.substring(0,o))+r.substring(s+1):o=s}}return r===``?null:r}function gi(e,t,n,r,i,a){var o=e.__className;if(D||o!==n||o===void 0){var s=hi(n,r,a);(!D||s!==e.getAttribute(`class`))&&(s==null?e.removeAttribute(`class`):t?e.className=s:e.setAttribute(`class`,s)),e.__className=n}else if(a&&i!==a)for(var c in a){var l=!!a[c];(i==null||l!==!!i[c])&&e.classList.toggle(c,l)}return a}var _i=Symbol(`is custom element`),vi=Symbol(`is html`),yi=Be?`link`:`LINK`;function bi(e,t,n,r){var i=xi(e);D&&(i[t]=e.getAttribute(t),t===`src`||t===`srcset`||t===`href`&&e.nodeName===yi)||i[t]!==(i[t]=n)&&(t===`loading`&&(e[Re]=n),n==null?e.removeAttribute(t):typeof n!=`string`&&Ci(e).includes(t)?e[t]=n:e.setAttribute(t,n))}function xi(e){return e.__attributes??={[_i]:e.nodeName.includes(`-`),[vi]:e.namespaceURI===tt}}var Si=new Map;function Ci(e){var t=e.getAttribute(`is`)||e.nodeName,n=Si.get(t);if(n)return n;Si.set(t,n=[]);for(var r,i=e,a=Element.prototype;a!==i;){for(var o in r=me(i),r)r[o].set&&n.push(o);i=_e(i)}return n}var wi={get(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(ye(r)&&(r=r()),typeof r==`object`&&r&&t in r)return r[t]}},set(e,t,n){let r=e.props.length;for(;r--;){let i=e.props[r];ye(i)&&(i=i());let a=pe(i,t);if(a&&a.set)return a.set(n),!0}return!1},getOwnPropertyDescriptor(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(ye(r)&&(r=r()),typeof r==`object`&&r&&t in r){let e=pe(r,t);return e&&!e.configurable&&(e.configurable=!0),e}}},has(e,t){if(t===Ie||t===Le)return!1;for(let n of e.props)if(ye(n)&&(n=n()),n!=null&&t in n)return!0;return!1},ownKeys(e){let t=[];for(let n of e.props)if(ye(n)&&(n=n()),n){for(let e in n)t.includes(e)||t.push(e);for(let e of Object.getOwnPropertySymbols(n))t.includes(e)||t.push(e)}return t}};function Ti(...e){return new Proxy({props:e},wi)}function Ei(e,t,n,r){var i=!pt||(n&2)!=0,a=(n&8)!=0,o=(n&16)!=0,s=r,c=!0,l=()=>(c&&(c=!1,s=o?Pr(r):r),s);let u;if(a){var d=Ie in e||Le in e;u=pe(e,t)?.set??(d&&t in e?n=>e[t]=n:void 0)}var f,p=!1;a?[f,p]=Bt(()=>e[t]):f=e[t],f===void 0&&r!==void 0&&(f=l(),u&&(i&&Je(t),u(f)));var m=i?()=>{var n=e[t];return n===void 0?l():(c=!0,n)}:()=>{var n=e[t];return n!==void 0&&(s=void 0),n===void 0?s:n};if(i&&!(n&4))return m;if(u){var h=e.$$legacy;return(function(e,t){return arguments.length>0?((!i||!t||h||p)&&u(t?m():e),e):m()})}var g=!1,_=(n&1?fn:mn)(()=>(g=!1,m()));a&&X(_);var v=G;return(function(e,t){if(arguments.length>0){let n=t?X(_):i&&a?An(e):e;return R(_,n),g=!0,s!==void 0&&(s=n),e}return gr&&g||v.f&16384?_.v:X(_)})}typeof window<`u`&&((window.__svelte??={}).v??=new Set).add(`5`);function Di(t){return e.Platform.isMacOS?t.metaKey:t.ctrlKey}var Oi=class{store;plugin;constructor(e,t){this.plugin=t,this.store=Nt(0),t.app.workspace.onLayoutReady(()=>{let{vault:n,metadataCache:r,workspace:i}=t.app;e.registerEvent(n.on(`create`,this.bump,this)),e.registerEvent(n.on(`delete`,this.bump,this)),e.registerEvent(n.on(`rename`,this.onRename,this)),e.registerEvent(r.on(`changed`,this.bump,this)),e.registerEvent(i.on(`periodic-notes:resolve`,this.bumpUnconditionally,this)),e.registerEvent(i.on(`periodic-notes:settings-updated`,this.bumpUnconditionally,this)),this.bump()})}bump(e){e&&!this.plugin.isPeriodic(e.path)||this.store.update(e=>e+1)}bumpUnconditionally(){this.store.update(e=>e+1)}onRename(e,t){(this.plugin.isPeriodic(e.path)||this.plugin.isPeriodic(t))&&this.store.update(e=>e+1)}getFile(e,t){return this.plugin.getPeriodicNote(t,e)}isGranularityEnabled(e){return this.plugin.settings.granularities[e]?.enabled??e===`day`}getEnabledGranularities(){return[`week`,`month`,`year`].filter(e=>this.plugin.settings.granularities[e]?.enabled)}};function ki(e,t){return`${e}:${t.format(n[e])}`}function Ai(e,t,n){let r=new Map,i=e[1].days[0];for(let i of e){for(let e of i.days)r.set(ki(`day`,e),t(e,`day`));if(n.includes(`week`)){let e=i.days[0];r.set(ki(`week`,e),t(e,`week`))}}return n.includes(`month`)&&r.set(ki(`month`,i),t(i,`month`)),n.includes(`year`)&&r.set(ki(`year`,i),t(i,`year`)),r}var ji=Q(`<td><div role="button" tabindex="0"> </div></td>`),Mi={hash:`svelte-1l5s8qw`,code:`.day.svelte-1l5s8qw {background-color:var(--color-background-day);border-radius:4px;color:var(--color-text-day);cursor:pointer;font-size:0.8em;height:100%;padding:4px;position:relative;text-align:center;transition:background-color 0.1s ease-in,
+let e=require(`obsidian`);function t(e,t){return`${e}:${t.clone().startOf(e).toISOString()}`}function n(e,t,n){let r=0,i=e.length-1;for(;r<=i;){let a=r+i>>>1,o=e[a];if(o===t)return e[a+(n===`forwards`?1:-1)]??null;o<t?r=a+1:i=a-1}return null}var r={day:`YYYY-MM-DD`,week:`gggg-[W]ww`,month:`YYYY-MM`,year:`YYYY`},i={enabled:!1,format:``,folder:``,templatePath:void 0},a={granularities:{day:{...i},week:{...i},month:{...i},year:{...i}}},o=[`sunday`,`monday`,`tuesday`,`wednesday`,`thursday`,`friday`,`saturday`],s=`calendar`,c=Symbol(`displayedMonth`),l=[`day`,`week`,`month`,`year`];function u(e,t){return e.granularities[t].format||r[t]}function d(e,t){let n=e.granularities[t].format;if(!n)return[r[t]];let i=/[^/]*$/.exec(n);return i?[n,i[0]]:[n]}function f(e,t){return e.granularities[t]}function p(e){return l.filter(t=>e.granularities[t].enabled)}function m(e){return e.replace(/\[[^\]]*\]/g,``).replace(/\\./g,``)}function h(e){return e.indexOf(`/`)===-1?e:e.split(`/`).pop()??``}function g(e){return!/[?<>\\:*|"]/g.test(e)&&!/[\x00-\x1f\x80-\x9f]/g.test(e)&&!/^\.+$/.test(e)&&!/^(con|prn|aux|nul|com[0-9]|lpt[0-9])(\..*)?$/i.test(e)}function _(e,t){if(!e)return``;if(!g(e))return`Format contains illegal characters`;if(t===`day`){let t=window.moment().format(e);if(!window.moment(t,e,!0).isValid())return`Failed to parse format`}return``}function v(e){let t=h(e).replace(/\[[^\]]*\]/g,``);return![`M`,`D`].every(e=>t.includes(e))||!(t.includes(`Y`)||t.includes(`y`))}function y(e,t){let n=window.moment().format(e);return window.moment(n,e,!0).isValid()?m(e).includes(`/`)&&t===`day`&&v(e)?`fragile-basename`:`valid`:`loose-parsing`}function b(...e){let t=[];for(let n=0,r=e.length;n<r;n++)t=t.concat(e[n].split(`/`));let n=[];for(let e=0,r=t.length;e<r;e++){let r=t[e];!r||r===`.`||n.push(r)}return t[0]===``&&n.unshift(``),n.join(`/`)}var x=/{{\s*(date|time)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,S=/{{\s*(month)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,C=/{{\s*(year)\s*(([-+]\d+)([ymwdhs]))?\s*(:.+?)?}}/gi,ee=RegExp(`{{\\s*(${o.join(`|`)})\\s*:(.*?)}}`,`gi`);function te(){let{moment:e}=window,t=e.localeData().firstDayOfWeek(),n=[...o];for(;t;){let e=n.shift();e&&n.push(e),t--}return n}function ne(e){let t=te().indexOf(e.toLowerCase());return Math.max(0,t)}function re(e,t,n,r,i){let a=window.moment();return e.replace(n,(e,n,o,s,c,l)=>{let u=t.clone();return i&&u.startOf(i),u.set({hour:a.get(`hour`),minute:a.get(`minute`),second:a.get(`second`)}),o&&u.add(parseInt(s,10),c),l?u.format(l.substring(1).trim()):u.format(r)})}function ie(e,t,n,r,i){let a=i.replace(/{{\s*date\s*}}/gi,e).replace(/{{\s*time\s*}}/gi,window.moment().format(`HH:mm`)).replace(/{{\s*title\s*}}/gi,e);return t===`day`&&(a=a.replace(/{{\s*yesterday\s*}}/gi,n.clone().subtract(1,`day`).format(r)).replace(/{{\s*tomorrow\s*}}/gi,n.clone().add(1,`d`).format(r)),a=re(a,n,x,r)),t===`week`&&(a=a.replace(ee,(e,t,r)=>{let i=ne(t);return n.weekday(i).format(r.trim())})),t===`month`&&(a=re(a,n,S,r,`month`)),t===`year`&&(a=re(a,n,C,r,`year`)),a}async function ae(t,n,r){if(!n||n===`/`)return``;let{metadataCache:i,vault:a}=t,o=(0,e.normalizePath)(n);try{let e=i.getFirstLinkpathDest(o,``);return e?a.cachedRead(e):``}catch(t){return console.error(`[Periodic Notes] Failed to read the ${r} note template '${o}'`,t),new e.Notice(`Failed to read the ${r} note template`),``}}async function oe(e,t,n,r){let i=u(n,r.granularity),a=await ae(e,n.granularities[r.granularity].templatePath,r.granularity),o=ie(t.basename,r.granularity,r.date,i,a);await e.vault.modify(t,o)}async function se(t,n,r){let i=(0,e.normalizePath)(b(r.folder??``,n.endsWith(`.md`)?n:`${n}.md`));return await ce(t,i),i}async function ce(e,t){let n=t.replace(/\\/g,`/`).split(`/`);n.pop();let r=``;for(let t of n)r=r?`${r}/${t}`:t,e.vault.getAbstractFileByPath(r)||await e.vault.createFolder(r)}function le(e){let t=e.extension.length+1;return e.path.slice(0,-t)}function ue(e,t,n){if(y(t,n)===`fragile-basename`){let n=le(e),r=(m(t).match(/\//g)?.length??0)+1;return n.split(`/`).slice(-r).join(`/`)}return e.basename}var de=class extends e.Component{byPath;byKey;sortedByGranularity;dirtyGranularities;constructor(t,n){super(),this.app=t,this.plugin=n,this.byPath=new Map,this.byKey=new Map,this.sortedByGranularity=new Map,this.dirtyGranularities=new Set(l),this.app.workspace.onLayoutReady(()=>{console.info(`[Periodic Notes] initializing cache`),this.initialize(),this.registerEvent(this.app.vault.on(`create`,t=>{t instanceof e.TFile&&this.resolve(t,`create`)})),this.registerEvent(this.app.vault.on(`delete`,t=>{t instanceof e.TFile&&this.remove(t.path)})),this.registerEvent(this.app.vault.on(`rename`,this.onRename,this)),this.registerEvent(this.app.metadataCache.on(`changed`,this.onMetadataChanged,this)),this.registerEvent(this.app.workspace.on(`periodic-notes:settings-updated`,this.reset,this))})}reset(){console.info(`[Periodic Notes] resetting cache`),this.byPath.clear(),this.byKey.clear(),this.sortedByGranularity.clear(),this.dirtyGranularities=new Set(l),this.initialize()}initialize(){let t=this.plugin.settings,n=new Set,r=(t,i)=>{if(!n.has(t)){n.add(t);for(let n of t.children)n instanceof e.TFile?i(n):n instanceof e.TFolder&&r(n,i)}},i=p(t);for(let n of i){let i=t.granularities[n].folder||`/`,a=this.app.vault.getAbstractFileByPath(i);a instanceof e.TFolder&&r(a,t=>{if(t instanceof e.TFile){this.resolve(t,`initialize`);let e=this.app.metadataCache.getFileCache(t);e&&this.onMetadataChanged(t,``,e)}})}}onMetadataChanged(t,n,r){let i=this.plugin.settings,a=p(i);if(a.length!==0)for(let n of a){let a=i.granularities[n].folder||`/`;if(!t.path.startsWith(a===`/`?``:`${a}/`))continue;let o=(0,e.parseFrontMatterEntry)(r.frontmatter,n);if(!o)continue;let s=u(i,n);if(typeof o==`string`){let e=window.moment(o,s,!0);e.isValid()&&this.set({filePath:t.path,date:e,granularity:n,match:`frontmatter`});return}}}onRename(t,n){t instanceof e.TFile&&(this.remove(n),this.resolve(t,`rename`))}resolve(t,n=`create`){let r=this.plugin.settings,i=p(r);if(i.length===0)return;let a=this.byPath.get(t.path);if(!(a&&a.match===`frontmatter`))for(let a of i){let i=r.granularities[a].folder||`/`;if(!t.path.startsWith(i===`/`?``:`${i}/`))continue;let o=d(r,a),s=ue(t,o[0],a),c=window.moment(s,o,!0);if(c.isValid()){let i={filePath:t.path,date:c,granularity:a,match:`filename`};this.set(i),n===`create`&&t.stat.size===0&&oe(this.app,t,r,i).catch(n=>{console.error(`[Periodic Notes] failed to apply template`,n),new e.Notice(`Periodic Notes: failed to apply template to "${t.path}". See console for details.`)}),this.app.workspace.trigger(`periodic-notes:resolve`,a,t);return}}}set(e){let n=t(e.granularity,e.date),r=this.byPath.get(e.filePath);if(r){let e=t(r.granularity,r.date);e!==n&&(this.byKey.delete(e),this.dirtyGranularities.add(r.granularity))}let i=this.byKey.get(n);i&&i.filePath!==e.filePath&&this.byPath.delete(i.filePath);let a=!this.byKey.has(n);this.byPath.set(e.filePath,e),this.byKey.set(n,e),a&&this.dirtyGranularities.add(e.granularity)}remove(e){let n=this.byPath.get(e);n&&(this.byKey.delete(t(n.granularity,n.date)),this.byPath.delete(e),this.dirtyGranularities.add(n.granularity))}getPeriodicNote(n,r){let i=t(n,r),a=this.byKey.get(i);if(!a)return null;let o=this.app.vault.getAbstractFileByPath(a.filePath);return o instanceof e.TFile?o:(this.remove(a.filePath),null)}isPeriodic(e,t){let n=this.byPath.get(e);return n?t?t===n.granularity:!0:!1}find(e){return e?this.byPath.get(e)??null:null}findAdjacent(e,r){let i=this.find(e);if(!i)return null;let a=n(this.getSortedKeys(i.granularity),t(i.granularity,i.date),r);return a?this.byKey.get(a)??null:null}getSortedKeys(e){if(!this.dirtyGranularities.has(e)){let t=this.sortedByGranularity.get(e);if(t)return t}let t=`${e}:`,n=[];for(let e of this.byKey.keys())e.startsWith(t)&&n.push(e);return n.sort(),this.sortedByGranularity.set(e,n),this.dirtyGranularities.delete(e),n}},fe=Array.isArray,pe=Array.prototype.indexOf,me=Array.prototype.includes,he=Array.from,ge=Object.defineProperty,_e=Object.getOwnPropertyDescriptor,ve=Object.getOwnPropertyDescriptors,ye=Object.prototype,be=Array.prototype,xe=Object.getPrototypeOf,Se=Object.isExtensible;function Ce(e){return typeof e==`function`}var we=()=>{};function Te(e){for(var t=0;t<e.length;t++)e[t]()}function Ee(){var e,t;return{promise:new Promise((n,r)=>{e=n,t=r}),resolve:e,reject:t}}var w=1024,T=2048,De=4096,Oe=8192,ke=16384,Ae=32768,je=1<<25,Me=65536,Ne=1<<19,Pe=1<<20,Fe=1<<25,Ie=65536,Le=1<<21,Re=1<<22,ze=1<<23,Be=Symbol(`$state`),Ve=Symbol(`legacy props`),He=Symbol(``),Ue=new class extends Error{name=`StaleReactionError`;message="The reaction that called `getAbortSignal()` was re-run or destroyed"},We=!!globalThis.document?.contentType&&globalThis.document.contentType.includes(`xml`);function Ge(e){throw Error(`https://svelte.dev/e/lifecycle_outside_component`)}function Ke(){throw Error(`https://svelte.dev/e/async_derived_orphan`)}function qe(e,t,n){throw Error(`https://svelte.dev/e/each_key_duplicate`)}function Je(e){throw Error(`https://svelte.dev/e/effect_in_teardown`)}function Ye(){throw Error(`https://svelte.dev/e/effect_in_unowned_derived`)}function Xe(e){throw Error(`https://svelte.dev/e/effect_orphan`)}function Ze(){throw Error(`https://svelte.dev/e/effect_update_depth_exceeded`)}function Qe(e){throw Error(`https://svelte.dev/e/props_invalid_value`)}function $e(){throw Error(`https://svelte.dev/e/set_context_after_init`)}function et(){throw Error(`https://svelte.dev/e/state_descriptors_fixed`)}function tt(){throw Error(`https://svelte.dev/e/state_prototype_fixed`)}function nt(){throw Error(`https://svelte.dev/e/state_unsafe_mutation`)}function rt(){throw Error(`https://svelte.dev/e/svelte_boundary_reset_onerror`)}var it={},E=Symbol(),at=`http://www.w3.org/1999/xhtml`;function ot(e){console.warn(`https://svelte.dev/e/hydration_mismatch`)}function st(){console.warn(`https://svelte.dev/e/svelte_boundary_reset_noop`)}var D=!1;function ct(e){D=e}var O;function k(e){if(e===null)throw ot(),it;return O=e}function lt(){return k(Hn(O))}function A(e){if(D){if(Hn(O)!==null)throw ot(),it;O=e}}function ut(e=1){if(D){for(var t=e,n=O;t--;)n=Hn(n);O=n}}function dt(e=!0){for(var t=0,n=O;;){if(n.nodeType===8){var r=n.data;if(r===`]`){if(t===0)return n;--t}else (r===`[`||r===`[!`||r[0]===`[`&&!isNaN(Number(r.slice(1))))&&(t+=1)}var i=Hn(n);e&&n.remove(),n=i}}function ft(e){if(!e||e.nodeType!==8)throw ot(),it;return e.data}function pt(e){return e===this.v}function mt(e,t){return e==e?e!==t||typeof e==`object`&&!!e||typeof e==`function`:t==t}function ht(e){return!mt(e,this.v)}var gt=!1,_t=!1,j=null;function vt(e){j=e}function yt(e){return wt(`getContext`).get(e)}function bt(e,t){let n=wt(`setContext`);if(gt){var r=G.f;!H&&r&32&&!j.i||$e()}return n.set(e,t),t}function xt(e,t=!1,n){j={p:j,i:!1,c:null,e:null,s:e,x:null,r:G,l:_t&&!t?{s:null,u:null,$:[]}:null}}function St(e){var t=j,n=t.e;if(n!==null){t.e=null;for(var r of n)tr(r)}return e!==void 0&&(t.x=e),t.i=!0,j=t.p,e??{}}function Ct(){return!_t||j!==null&&j.l===null}function wt(e){return j===null&&Ge(e),j.c??=new Map(Tt(j)||void 0)}function Tt(e){let t=e.p;for(;t!==null;){let e=t.c;if(e!==null)return e;t=t.p}return null}var Et=[];function Dt(){var e=Et;Et=[],Te(e)}function Ot(e){if(Et.length===0&&!Kt){var t=Et;queueMicrotask(()=>{t===Et&&Dt()})}Et.push(e)}function kt(e){var t=G;if(t===null)return H.f|=ze,e;if(!(t.f&32768)&&!(t.f&4))throw e;At(e,t)}function At(e,t){for(;t!==null;){if(t.f&128){if(!(t.f&32768))throw e;try{t.b.error(e);return}catch(t){e=t}}t=t.parent}throw e}var jt=~(T|De|w);function M(e,t){e.f=e.f&jt|t}function Mt(e){e.f&512||e.deps===null?M(e,w):M(e,De)}function Nt(e){if(e!==null)for(let t of e)!(t.f&2)||!(t.f&65536)||(t.f^=Ie,Nt(t.deps))}function Pt(e,t,n){e.f&2048?t.add(e):e.f&4096&&n.add(e),Nt(e.deps),M(e,w)}function Ft(e,t,n){if(e==null)return t(void 0),n&&n(void 0),we;let r=Rr(()=>e.subscribe(t,n));return r.unsubscribe?()=>r.unsubscribe():r}var It=[];function Lt(e,t=we){let n=null,r=new Set;function i(t){if(mt(e,t)&&(e=t,n)){let t=!It.length;for(let t of r)t[1](),It.push(t,e);if(t){for(let e=0;e<It.length;e+=2)It[e][0](It[e+1]);It.length=0}}}function a(t){i(t(e))}function o(o,s=we){let c=[o,s];return r.add(c),r.size===1&&(n=t(i,a)||we),o(e),()=>{r.delete(c),r.size===0&&n&&(n(),n=null)}}return{set:i,update:a,subscribe:o}}function Rt(e){let t;return Ft(e,e=>t=e)(),t}var zt=!1,Bt=!1,Vt=Symbol();function Ht(e,t,n){let r=n[t]??={store:null,source:kn(void 0),unsubscribe:we};if(r.store!==e&&!(Vt in n))if(r.unsubscribe(),r.store=e??null,e==null)r.source.v=void 0,r.unsubscribe=we;else{var i=!0;r.unsubscribe=Ft(e,e=>{i?r.source.v=e:R(r.source,e)}),i=!1}return e&&Vt in n?Rt(e):X(r.source)}function Ut(){let e={};function t(){$n(()=>{for(var t in e)e[t].unsubscribe();ge(e,Vt,{enumerable:!1,value:!0})})}return[e,t]}function Wt(e){var t=Bt;try{return Bt=!1,[e(),Bt]}finally{Bt=t}}var N=new Set,P=null,F=null,Gt=null,Kt=!1,qt=!1,Jt=null,Yt=null,Xt=0,Zt=1,Qt=class e{id=Zt++;current=new Map;previous=new Map;#e=new Set;#t=new Set;#n=new Map;#r=new Map;#i=null;#a=[];#o=[];#s=new Set;#c=new Set;#l=new Map;is_fork=!1;#u=!1;#d=new Set;#f(){return this.is_fork||this.#r.size>0}#p(){for(let n of this.#d)for(let r of n.#r.keys()){for(var e=!1,t=r;t.parent!==null;){if(this.#l.has(t)){e=!0;break}t=t.parent}if(!e)return!0}return!1}skip_effect(e){this.#l.has(e)||this.#l.set(e,{d:[],m:[]})}unskip_effect(e){var t=this.#l.get(e);if(t){this.#l.delete(e);for(var n of t.d)M(n,T),this.schedule(n);for(n of t.m)M(n,De),this.schedule(n)}}#m(){if(Xt++>1e3&&(N.delete(this),$t()),!this.#f()){for(let e of this.#s)this.#c.delete(e),M(e,T),this.schedule(e);for(let e of this.#c)M(e,De),this.schedule(e)}let t=this.#a;this.#a=[],this.apply();var n=Jt=[],r=[],i=Yt=[];for(let e of t)try{this.#h(e,n,r)}catch(t){throw sn(e),t}if(P=null,i.length>0){var a=e.ensure();for(let e of i)a.schedule(e)}if(Jt=null,Yt=null,this.#f()||this.#p()){this.#g(r),this.#g(n);for(let[e,t]of this.#l)on(e,t)}else{this.#n.size===0&&N.delete(this),this.#s.clear(),this.#c.clear();for(let e of this.#e)e(this);this.#e.clear(),tn(r),tn(n),this.#i?.resolve()}var o=P;if(this.#a.length>0){let e=o??=this;e.#a.push(...this.#a.filter(t=>!e.#a.includes(t)))}o!==null&&(N.add(o),o.#m()),N.has(this)||this.#_()}#h(e,t,n){e.f^=w;for(var r=e.first;r!==null;){var i=r.f,a=(i&96)!=0;if(!(a&&i&1024||i&8192||this.#l.has(r))&&r.fn!==null){a?r.f^=w:i&4?t.push(r):gt&&i&16777224?n.push(r):Ar(r)&&(i&16&&this.#c.add(r),Fr(r));var o=r.first;if(o!==null){r=o;continue}}for(;r!==null;){var s=r.next;if(s!==null){r=s;break}r=r.parent}}}#g(e){for(var t=0;t<e.length;t+=1)Pt(e[t],this.#s,this.#c)}capture(e,t,n=!1){t!==E&&!this.previous.has(e)&&this.previous.set(e,t),e.f&8388608||(this.current.set(e,[e.v,n]),F?.set(e,e.v))}activate(){P=this}deactivate(){P=null,F=null}flush(){try{qt=!0,P=this,this.#m()}finally{Xt=0,Gt=null,Jt=null,Yt=null,qt=!1,P=null,F=null,En.clear()}}discard(){for(let e of this.#t)e(this);this.#t.clear(),N.delete(this)}register_created_effect(e){this.#o.push(e)}#_(){for(let l of N){var e=l.id<this.id,t=[];for(let[r,[i,a]]of this.current){if(l.current.has(r)){var n=l.current.get(r)[0];if(e&&i!==n)l.current.set(r,[i,a]);else continue}t.push(r)}var r=[...l.current.keys()].filter(e=>!this.current.has(e));if(r.length===0)e&&l.discard();else if(t.length>0){l.activate();var i=new Set,a=new Map;for(var o of t)nn(o,r,i,a);a=new Map;var s=[...l.current.keys()].filter(e=>this.current.has(e)?this.current.get(e)[0]!==e:!0);for(let e of this.#o)!(e.f&155648)&&rn(e,s,a)&&(e.f&4194320?(M(e,T),l.schedule(e)):l.#s.add(e));if(l.#a.length>0){l.apply();for(var c of l.#a)l.#h(c,[],[]);l.#a=[]}l.deactivate()}}for(let e of N)e.#d.has(this)&&(e.#d.delete(this),e.#d.size===0&&!e.#f()&&(e.activate(),e.#m()))}increment(e,t){let n=this.#n.get(t)??0;if(this.#n.set(t,n+1),e){let e=this.#r.get(t)??0;this.#r.set(t,e+1)}}decrement(e,t,n){let r=this.#n.get(t)??0;if(r===1?this.#n.delete(t):this.#n.set(t,r-1),e){let e=this.#r.get(t)??0;e===1?this.#r.delete(t):this.#r.set(t,e-1)}this.#u||n||(this.#u=!0,Ot(()=>{this.#u=!1,this.flush()}))}transfer_effects(e,t){for(let t of e)this.#s.add(t);for(let e of t)this.#c.add(e);e.clear(),t.clear()}oncommit(e){this.#e.add(e)}ondiscard(e){this.#t.add(e)}settled(){return(this.#i??=Ee()).promise}static ensure(){if(P===null){let t=P=new e;qt||(N.add(P),Kt||Ot(()=>{P===t&&t.flush()}))}return P}apply(){if(!gt||!this.is_fork&&N.size===1){F=null;return}F=new Map;for(let[e,[t]]of this.current)F.set(e,t);for(let n of N)if(!(n===this||n.is_fork)){var e=!1,t=!1;if(n.id<this.id)for(let[r,[,i]]of n.current)i||(e||=this.current.has(r),t||=!this.current.has(r));if(e&&t)this.#d.add(n);else for(let[e,t]of n.previous)F.has(e)||F.set(e,t)}}schedule(e){if(Gt=e,e.b?.is_pending&&e.f&16777228&&!(e.f&32768)){e.b.defer_effect(e);return}for(var t=e;t.parent!==null;){t=t.parent;var n=t.f;if(Jt!==null&&t===G&&(gt||(H===null||!(H.f&2))&&!zt))return;if(n&96){if(!(n&1024))return;t.f^=w}}this.#a.push(t)}};function $t(){try{Ze()}catch(e){At(e,Gt)}}var en=null;function tn(e){var t=e.length;if(t!==0){for(var n=0;n<t;){var r=e[n++];if(!(r.f&24576)&&Ar(r)&&(en=new Set,Fr(r),r.deps===null&&r.first===null&&r.nodes===null&&r.teardown===null&&r.ac===null&&fr(r),en?.size>0)){En.clear();for(let e of en){if(e.f&24576)continue;let t=[e],n=e.parent;for(;n!==null;)en.has(n)&&(en.delete(n),t.push(n)),n=n.parent;for(let e=t.length-1;e>=0;e--){let n=t[e];n.f&24576||Fr(n)}}en.clear()}}en=null}}function nn(e,t,n,r){if(!n.has(e)&&(n.add(e),e.reactions!==null))for(let i of e.reactions){let e=i.f;e&2?nn(i,t,n,r):e&4194320&&!(e&2048)&&rn(i,t,r)&&(M(i,T),an(i))}}function rn(e,t,n){let r=n.get(e);if(r!==void 0)return r;if(e.deps!==null)for(let r of e.deps){if(me.call(t,r))return!0;if(r.f&2&&rn(r,t,n))return n.set(r,!0),!0}return n.set(e,!1),!1}function an(e){P.schedule(e)}function on(e,t){if(!(e.f&32&&e.f&1024)){e.f&2048?t.d.push(e):e.f&4096&&t.m.push(e),M(e,w);for(var n=e.first;n!==null;)on(n,t),n=n.next}}function sn(e){M(e,w);for(var t=e.first;t!==null;)sn(t),t=t.next}function cn(e){let t=0,n=On(0),r;return()=>{Qn()&&(X(n),ar(()=>(t===0&&(r=Rr(()=>e(()=>Mn(n)))),t+=1,()=>{Ot(()=>{--t,t===0&&(r?.(),r=void 0,Mn(n))})})))}}var ln=Me|Ne;function un(e,t,n,r){new dn(e,t,n,r)}var dn=class{parent;is_pending=!1;transform_error;#e;#t=D?O:null;#n;#r;#i;#a=null;#o=null;#s=null;#c=null;#l=0;#u=0;#d=!1;#f=new Set;#p=new Set;#m=null;#h=cn(()=>(this.#m=On(this.#l),()=>{this.#m=null}));constructor(e,t,n,r){this.#e=e,this.#n=t,this.#r=e=>{var t=G;t.b=this,t.f|=128,n(e)},this.parent=G.b,this.transform_error=r??this.parent?.transform_error??(e=>e),this.#i=sr(()=>{if(D){let e=this.#t;lt();let t=e.data===`[!`;if(e.data.startsWith(`[?`)){let t=JSON.parse(e.data.slice(2));this.#_(t)}else t?this.#v():this.#g()}else this.#y()},ln),D&&(this.#e=O)}#g(){try{this.#a=B(()=>this.#r(this.#e))}catch(e){this.error(e)}}#_(e){let t=this.#n.failed;t&&(this.#s=B(()=>{t(this.#e,()=>e,()=>()=>{})}))}#v(){let e=this.#n.pending;e&&(this.is_pending=!0,this.#o=B(()=>e(this.#e)),Ot(()=>{var e=this.#c=document.createDocumentFragment(),t=Bn();e.append(t),this.#a=this.#x(()=>B(()=>this.#r(t))),this.#u===0&&(this.#e.before(e),this.#c=null,pr(this.#o,()=>{this.#o=null}),this.#b(P))}))}#y(){try{if(this.is_pending=this.has_pending_snippet(),this.#u=0,this.#l=0,this.#a=B(()=>{this.#r(this.#e)}),this.#u>0){var e=this.#c=document.createDocumentFragment();_r(this.#a,e);let t=this.#n.pending;this.#o=B(()=>t(this.#e))}else this.#b(P)}catch(e){this.error(e)}}#b(e){this.is_pending=!1,e.transfer_effects(this.#f,this.#p)}defer_effect(e){Pt(e,this.#f,this.#p)}is_rendered(){return!this.is_pending&&(!this.parent||this.parent.is_rendered())}has_pending_snippet(){return!!this.#n.pending}#x(e){var t=G,n=H,r=j;Sr(this.#i),W(this.#i),vt(this.#i.ctx);try{return Qt.ensure(),e()}catch(e){return kt(e),null}finally{Sr(t),W(n),vt(r)}}#S(e,t){if(!this.has_pending_snippet()){this.parent&&this.parent.#S(e,t);return}this.#u+=e,this.#u===0&&(this.#b(t),this.#o&&pr(this.#o,()=>{this.#o=null}),this.#c&&=(this.#e.before(this.#c),null))}update_pending_count(e,t){this.#S(e,t),this.#l+=e,!(!this.#m||this.#d)&&(this.#d=!0,Ot(()=>{this.#d=!1,this.#m&&An(this.#m,this.#l)}))}get_effect_pending(){return this.#h(),X(this.#m)}error(e){var t=this.#n.onerror;let n=this.#n.failed;if(!t&&!n)throw e;this.#a&&=(V(this.#a),null),this.#o&&=(V(this.#o),null),this.#s&&=(V(this.#s),null),D&&(k(this.#t),ut(),k(dt()));var r=!1,i=!1;let a=()=>{if(r){st();return}r=!0,i&&rt(),this.#s!==null&&pr(this.#s,()=>{this.#s=null}),this.#x(()=>{this.#y()})},o=e=>{try{i=!0,t?.(e,a),i=!1}catch(e){At(e,this.#i&&this.#i.parent)}n&&(this.#s=this.#x(()=>{try{return B(()=>{var t=G;t.b=this,t.f|=128,n(this.#e,()=>e,()=>a)})}catch(e){return At(e,this.#i.parent),null}}))};Ot(()=>{var t;try{t=this.transform_error(e)}catch(e){At(e,this.#i&&this.#i.parent);return}typeof t==`object`&&t&&typeof t.then==`function`?t.then(o,e=>At(e,this.#i&&this.#i.parent)):o(t)})}};function fn(e,t,n,r){let i=Ct()?gn:vn;var a=e.filter(e=>!e.settled);if(n.length===0&&a.length===0){r(t.map(i));return}var o=G,s=pn(),c=a.length===1?a[0].promise:a.length>1?Promise.all(a.map(e=>e.promise)):null;function l(e){s();try{r(e)}catch(e){o.f&16384||At(e,o)}mn()}if(n.length===0){c.then(()=>l(t.map(i)));return}var u=hn();function d(){Promise.all(n.map(e=>_n(e))).then(e=>l([...t.map(i),...e])).catch(e=>At(e,o)).finally(()=>u())}c?c.then(()=>{s(),d(),mn()}):d()}function pn(){var e=G,t=H,n=j,r=P;return function(i=!0){Sr(e),W(t),vt(n),i&&!(e.f&16384)&&(r?.activate(),r?.apply())}}function mn(e=!0){Sr(null),W(null),vt(null),e&&P?.deactivate()}function hn(){var e=G,t=e.b,n=P,r=t.is_rendered();return t.update_pending_count(1,n),n.increment(r,e),(i=!1)=>{t.update_pending_count(-1,n),n.decrement(r,e,i)}}function gn(e){var t=2|T,n=H!==null&&H.f&2?H:null;return G!==null&&(G.f|=Ne),{ctx:j,deps:null,effects:null,equals:pt,f:t,fn:e,reactions:null,rv:0,v:E,wv:0,parent:n??G,ac:null}}function _n(e,t,n){let r=G;r===null&&Ke();var i=void 0,a=On(E),o=!H,s=new Map;return ir(()=>{var t=G,n=Ee();i=n.promise;try{Promise.resolve(e()).then(n.resolve,n.reject).finally(mn)}catch(e){n.reject(e),mn()}var c=P;if(o){if(t.f&32768)var l=hn();if(r.b.is_rendered())s.get(c)?.reject(Ue),s.delete(c);else{for(let e of s.values())e.reject(Ue);s.clear()}s.set(c,n)}let u=(e,n=void 0)=>{if(l&&l(n===Ue),!(n===Ue||t.f&16384)){if(c.activate(),n)a.f|=ze,An(a,n);else{a.f&8388608&&(a.f^=ze),An(a,e);for(let[e,t]of s){if(s.delete(e),e===c)break;t.reject(Ue)}}c.deactivate()}};n.promise.then(u,e=>u(null,e||`unknown`))}),$n(()=>{for(let e of s.values())e.reject(Ue)}),new Promise(e=>{function t(n){function r(){n===i?e(a):t(i)}n.then(r,r)}t(i)})}function I(e){let t=gn(e);return gt||Cr(t),t}function vn(e){let t=gn(e);return t.equals=ht,t}function yn(e){var t=e.effects;if(t!==null){e.effects=null;for(var n=0;n<t.length;n+=1)V(t[n])}}function bn(e){for(var t=e.parent;t!==null;){if(!(t.f&2))return t.f&16384?null:t;t=t.parent}return null}function xn(e){var t,n=G;Sr(bn(e));try{e.f&=~Ie,yn(e),t=Mr(e)}finally{Sr(n)}return t}function Sn(e){var t=e.v,n=xn(e);if(!e.equals(n)&&(e.wv=kr(),(!P?.is_fork||e.deps===null)&&(e.v=n,P?.capture(e,t,!0),e.deps===null))){M(e,w);return}br||(F===null?Mt(e):(Qn()||P?.is_fork)&&F.set(e,n))}function Cn(e){if(e.effects!==null)for(let t of e.effects)(t.teardown||t.ac)&&(t.teardown?.(),t.ac?.abort(Ue),t.teardown=we,t.ac=null,Pr(t,0),lr(t))}function wn(e){if(e.effects!==null)for(let t of e.effects)t.teardown&&Fr(t)}var Tn=new Set,En=new Map,Dn=!1;function On(e,t){return{f:0,v:e,reactions:null,equals:pt,rv:0,wv:0}}function L(e,t){let n=On(e,t);return Cr(n),n}function kn(e,t=!1,n=!0){let r=On(e);return t||(r.equals=ht),_t&&n&&j!==null&&j.l!==null&&(j.l.s??=[]).push(r),r}function R(e,t,n=!1){return H!==null&&(!U||H.f&131072)&&Ct()&&H.f&4325394&&(K===null||!me.call(K,e))&&nt(),An(e,n?Pn(t):t,Yt)}function An(e,t,n=null){if(!e.equals(t)){var r=e.v;br?En.set(e,t):En.set(e,r),e.v=t;var i=Qt.ensure();if(i.capture(e,r),e.f&2){let t=e;e.f&2048&&xn(t),F===null&&Mt(t)}e.wv=kr(),Nn(e,T,n),Ct()&&G!==null&&G.f&1024&&!(G.f&96)&&(Y===null?wr([e]):Y.push(e)),!i.is_fork&&Tn.size>0&&!Dn&&jn()}return t}function jn(){Dn=!1;for(let e of Tn)e.f&1024&&M(e,De),Ar(e)&&Fr(e);Tn.clear()}function Mn(e){R(e,e.v+1)}function Nn(e,t,n){var r=e.reactions;if(r!==null)for(var i=Ct(),a=r.length,o=0;o<a;o++){var s=r[o],c=s.f;if(!(!i&&s===G)){var l=(c&T)===0;if(l&&M(s,t),c&2){var u=s;F?.delete(u),c&65536||(c&512&&(s.f|=Ie),Nn(u,De,n))}else if(l){var d=s;c&16&&en!==null&&en.add(d),n===null?an(d):n.push(d)}}}}function Pn(e){if(typeof e!=`object`||!e||Be in e)return e;let t=xe(e);if(t!==ye&&t!==be)return e;var n=new Map,r=fe(e),i=L(0),a=null,o=Dr,s=e=>{if(Dr===o)return e();var t=H,n=Dr;W(null),Or(o);var r=e();return W(t),Or(n),r};return r&&n.set(`length`,L(e.length,a)),new Proxy(e,{defineProperty(e,t,r){(!(`value`in r)||r.configurable===!1||r.enumerable===!1||r.writable===!1)&&et();var i=n.get(t);return i===void 0?s(()=>{var e=L(r.value,a);return n.set(t,e),e}):R(i,r.value,!0),!0},deleteProperty(e,t){var r=n.get(t);if(r===void 0){if(t in e){let e=s(()=>L(E,a));n.set(t,e),Mn(i)}}else R(r,E),Mn(i);return!0},get(t,r,i){if(r===Be)return e;var o=n.get(r),c=r in t;if(o===void 0&&(!c||_e(t,r)?.writable)&&(o=s(()=>L(Pn(c?t[r]:E),a)),n.set(r,o)),o!==void 0){var l=X(o);return l===E?void 0:l}return Reflect.get(t,r,i)},getOwnPropertyDescriptor(e,t){var r=Reflect.getOwnPropertyDescriptor(e,t);if(r&&`value`in r){var i=n.get(t);i&&(r.value=X(i))}else if(r===void 0){var a=n.get(t),o=a?.v;if(a!==void 0&&o!==E)return{enumerable:!0,configurable:!0,value:o,writable:!0}}return r},has(e,t){if(t===Be)return!0;var r=n.get(t),i=r!==void 0&&r.v!==E||Reflect.has(e,t);return(r!==void 0||G!==null&&(!i||_e(e,t)?.writable))&&(r===void 0&&(r=s(()=>L(i?Pn(e[t]):E,a)),n.set(t,r)),X(r)===E)?!1:i},set(e,t,o,c){var l=n.get(t),u=t in e;if(r&&t===`length`)for(var d=o;d<l.v;d+=1){var f=n.get(d+``);f===void 0?d in e&&(f=s(()=>L(E,a)),n.set(d+``,f)):R(f,E)}if(l===void 0)(!u||_e(e,t)?.writable)&&(l=s(()=>L(void 0,a)),R(l,Pn(o)),n.set(t,l));else{u=l.v!==E;var p=s(()=>Pn(o));R(l,p)}var m=Reflect.getOwnPropertyDescriptor(e,t);if(m?.set&&m.set.call(c,o),!u){if(r&&typeof t==`string`){var h=n.get(`length`),g=Number(t);Number.isInteger(g)&&g>=h.v&&R(h,g+1)}Mn(i)}return!0},ownKeys(e){X(i);var t=Reflect.ownKeys(e).filter(e=>{var t=n.get(e);return t===void 0||t.v!==E});for(var[r,a]of n)a.v!==E&&!(r in e)&&t.push(r);return t},setPrototypeOf(){tt()}})}var Fn,In,Ln,Rn;function zn(){if(Fn===void 0){Fn=window,In=/Firefox/.test(navigator.userAgent);var e=Element.prototype,t=Node.prototype,n=Text.prototype;Ln=_e(t,`firstChild`).get,Rn=_e(t,`nextSibling`).get,Se(e)&&(e.__click=void 0,e.__className=void 0,e.__attributes=null,e.__style=void 0,e.__e=void 0),Se(n)&&(n.__t=void 0)}}function Bn(e=``){return document.createTextNode(e)}function Vn(e){return Ln.call(e)}function Hn(e){return Rn.call(e)}function z(e,t){if(!D)return Vn(e);var n=Vn(O);if(n===null)n=O.appendChild(Bn());else if(t&&n.nodeType!==3){var r=Bn();return n?.before(r),k(r),r}return t&&qn(n),k(n),n}function Un(e,t=1,n=!1){let r=D?O:e;for(var i;t--;)i=r,r=Hn(r);if(!D)return r;if(n){if(r?.nodeType!==3){var a=Bn();return r===null?i?.after(a):r.before(a),k(a),a}qn(r)}return k(r),r}function Wn(e){e.textContent=``}function Gn(){return!gt||en!==null?!1:(G.f&Ae)!==0}function Kn(e,t,n){let r=n?{is:n}:void 0;return document.createElementNS(t??`http://www.w3.org/1999/xhtml`,e,r)}function qn(e){if(e.nodeValue.length<65536)return;let t=e.nextSibling;for(;t!==null&&t.nodeType===3;)t.remove(),e.nodeValue+=t.nodeValue,t=e.nextSibling}function Jn(e){var t=H,n=G;W(null),Sr(null);try{return e()}finally{W(t),Sr(n)}}function Yn(e){G===null&&(H===null&&Xe(e),Ye()),br&&Je(e)}function Xn(e,t){var n=t.last;n===null?t.last=t.first=e:(n.next=e,e.prev=n,t.last=e)}function Zn(e,t){var n=G;n!==null&&n.f&8192&&(e|=Oe);var r={ctx:j,deps:null,nodes:null,f:e|T|512,first:null,fn:t,last:null,next:null,parent:n,b:n&&n.b,prev:null,teardown:null,wv:0,ac:null};P?.register_created_effect(r);var i=r;if(e&4)Jt===null?Qt.ensure().schedule(r):Jt.push(r);else if(t!==null){try{Fr(r)}catch(e){throw V(r),e}i.deps===null&&i.teardown===null&&i.nodes===null&&i.first===i.last&&!(i.f&524288)&&(i=i.first,e&16&&e&65536&&i!==null&&(i.f|=Me))}if(i!==null&&(i.parent=n,n!==null&&Xn(i,n),H!==null&&H.f&2&&!(e&64))){var a=H;(a.effects??=[]).push(i)}return r}function Qn(){return H!==null&&!U}function $n(e){let t=Zn(8,null);return M(t,w),t.teardown=e,t}function er(e){Yn(`$effect`);var t=G.f;if(!H&&t&32&&!(t&32768)){var n=j;(n.e??=[]).push(e)}else return tr(e)}function tr(e){return Zn(4|Pe,e)}function nr(e){Qt.ensure();let t=Zn(64|Ne,e);return(e={})=>new Promise(n=>{e.outro?pr(t,()=>{V(t),n(void 0)}):(V(t),n(void 0))})}function rr(e){return Zn(4,e)}function ir(e){return Zn(Re|Ne,e)}function ar(e,t=0){return Zn(8|t,e)}function or(e,t=[],n=[],r=[]){fn(r,t,n,t=>{Zn(8,()=>e(...t.map(X)))})}function sr(e,t=0){return Zn(16|t,e)}function B(e){return Zn(32|Ne,e)}function cr(e){var t=e.teardown;if(t!==null){let e=br,n=H;xr(!0),W(null);try{t.call(null)}finally{xr(e),W(n)}}}function lr(e,t=!1){var n=e.first;for(e.first=e.last=null;n!==null;){let e=n.ac;e!==null&&Jn(()=>{e.abort(Ue)});var r=n.next;n.f&64?n.parent=null:V(n,t),n=r}}function ur(e){for(var t=e.first;t!==null;){var n=t.next;t.f&32||V(t),t=n}}function V(e,t=!0){var n=!1;(t||e.f&262144)&&e.nodes!==null&&e.nodes.end!==null&&(dr(e.nodes.start,e.nodes.end),n=!0),M(e,je),lr(e,t&&!n),Pr(e,0);var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)e.stop();cr(e),e.f^=je,e.f|=ke;var i=e.parent;i!==null&&i.first!==null&&fr(e),e.next=e.prev=e.teardown=e.ctx=e.deps=e.fn=e.nodes=e.ac=e.b=null}function dr(e,t){for(;e!==null;){var n=e===t?null:Hn(e);e.remove(),e=n}}function fr(e){var t=e.parent,n=e.prev,r=e.next;n!==null&&(n.next=r),r!==null&&(r.prev=n),t!==null&&(t.first===e&&(t.first=r),t.last===e&&(t.last=n))}function pr(e,t,n=!0){var r=[];mr(e,r,!0);var i=()=>{n&&V(e),t&&t()},a=r.length;if(a>0){var o=()=>--a||i();for(var s of r)s.out(o)}else i()}function mr(e,t,n){if(!(e.f&8192)){e.f^=Oe;var r=e.nodes&&e.nodes.t;if(r!==null)for(let e of r)(e.is_global||n)&&t.push(e);for(var i=e.first;i!==null;){var a=i.next,o=(i.f&65536)!=0||(i.f&32)!=0&&(e.f&16)!=0;mr(i,t,o?n:!1),i=a}}}function hr(e){gr(e,!0)}function gr(e,t){if(e.f&8192){e.f^=Oe,e.f&1024||(M(e,T),Qt.ensure().schedule(e));for(var n=e.first;n!==null;){var r=n.next,i=(n.f&65536)!=0||(n.f&32)!=0;gr(n,i?t:!1),n=r}var a=e.nodes&&e.nodes.t;if(a!==null)for(let e of a)(e.is_global||t)&&e.in()}}function _r(e,t){if(e.nodes)for(var n=e.nodes.start,r=e.nodes.end;n!==null;){var i=n===r?null:Hn(n);t.append(n),n=i}}var vr=null,yr=!1,br=!1;function xr(e){br=e}var H=null,U=!1;function W(e){H=e}var G=null;function Sr(e){G=e}var K=null;function Cr(e){H!==null&&(!gt||H.f&2)&&(K===null?K=[e]:K.push(e))}var q=null,J=0,Y=null;function wr(e){Y=e}var Tr=1,Er=0,Dr=Er;function Or(e){Dr=e}function kr(){return++Tr}function Ar(e){var t=e.f;if(t&2048)return!0;if(t&2&&(e.f&=~Ie),t&4096){for(var n=e.deps,r=n.length,i=0;i<r;i++){var a=n[i];if(Ar(a)&&Sn(a),a.wv>e.wv)return!0}t&512&&F===null&&M(e,w)}return!1}function jr(e,t,n=!0){var r=e.reactions;if(r!==null&&!(!gt&&K!==null&&me.call(K,e)))for(var i=0;i<r.length;i++){var a=r[i];a.f&2?jr(a,t,!1):t===a&&(n?M(a,T):a.f&1024&&M(a,De),an(a))}}function Mr(e){var t=q,n=J,r=Y,i=H,a=K,o=j,s=U,c=Dr,l=e.f;q=null,J=0,Y=null,H=l&96?null:e,K=null,vt(e.ctx),U=!1,Dr=++Er,e.ac!==null&&(Jn(()=>{e.ac.abort(Ue)}),e.ac=null);try{e.f|=Le;var u=e.fn,d=u();e.f|=Ae;var f=e.deps,p=P?.is_fork;if(q!==null){var m;if(p||Pr(e,J),f!==null&&J>0)for(f.length=J+q.length,m=0;m<q.length;m++)f[J+m]=q[m];else e.deps=f=q;if(Qn()&&e.f&512)for(m=J;m<f.length;m++)(f[m].reactions??=[]).push(e)}else !p&&f!==null&&J<f.length&&(Pr(e,J),f.length=J);if(Ct()&&Y!==null&&!U&&f!==null&&!(e.f&6146))for(m=0;m<Y.length;m++)jr(Y[m],e);if(i!==null&&i!==e){if(Er++,i.deps!==null)for(let e=0;e<n;e+=1)i.deps[e].rv=Er;if(t!==null)for(let e of t)e.rv=Er;Y!==null&&(r===null?r=Y:r.push(...Y))}return e.f&8388608&&(e.f^=ze),d}catch(e){return kt(e)}finally{e.f^=Le,q=t,J=n,Y=r,H=i,K=a,vt(o),U=s,Dr=c}}function Nr(e,t){let n=t.reactions;if(n!==null){var r=pe.call(n,e);if(r!==-1){var i=n.length-1;i===0?n=t.reactions=null:(n[r]=n[i],n.pop())}}if(n===null&&t.f&2&&(q===null||!me.call(q,t))){var a=t;a.f&512&&(a.f^=512,a.f&=~Ie),Mt(a),Cn(a),Pr(a,0)}}function Pr(e,t){var n=e.deps;if(n!==null)for(var r=t;r<n.length;r++)Nr(e,n[r])}function Fr(e){var t=e.f;if(!(t&16384)){M(e,w);var n=G,r=yr;G=e,yr=!0;try{t&16777232?ur(e):lr(e),cr(e);var i=Mr(e);e.teardown=typeof i==`function`?i:null,e.wv=Tr}finally{yr=r,G=n}}}function X(e){var t=(e.f&2)!=0;if(vr?.add(e),H!==null&&!U&&!(G!==null&&G.f&16384)&&(K===null||!me.call(K,e))){var n=H.deps;if(H.f&2097152)e.rv<Er&&(e.rv=Er,q===null&&n!==null&&n[J]===e?J++:q===null?q=[e]:q.push(e));else{(H.deps??=[]).push(e);var r=e.reactions;r===null?e.reactions=[H]:me.call(r,H)||r.push(H)}}if(br&&En.has(e))return En.get(e);if(t){var i=e;if(br){var a=i.v;return(!(i.f&1024)&&i.reactions!==null||Lr(i))&&(a=xn(i)),En.set(i,a),a}var o=(i.f&512)==0&&!U&&H!==null&&(yr||(H.f&512)!=0),s=(i.f&Ae)===0;Ar(i)&&(o&&(i.f|=512),Sn(i)),o&&!s&&(wn(i),Ir(i))}if(F?.has(e))return F.get(e);if(e.f&8388608)throw e.v;return e.v}function Ir(e){if(e.f|=512,e.deps!==null)for(let t of e.deps)(t.reactions??=[]).push(e),t.f&2&&!(t.f&512)&&(wn(t),Ir(t))}function Lr(e){if(e.v===E)return!0;if(e.deps===null)return!1;for(let t of e.deps)if(En.has(t)||t.f&2&&Lr(t))return!0;return!1}function Rr(e){var t=U;try{return U=!0,e()}finally{U=t}}[...`allowfullscreen.async.autofocus.autoplay.checked.controls.default.disabled.formnovalidate.indeterminate.inert.ismap.loop.multiple.muted.nomodule.novalidate.open.playsinline.readonly.required.reversed.seamless.selected.webkitdirectory.defer.disablepictureinpicture.disableremoteplayback`.split(`.`)];var zr=[`touchstart`,`touchmove`];function Br(e){return zr.includes(e)}var Vr=Symbol(`events`),Hr=new Set,Ur=new Set;function Wr(e,t,n,r={}){function i(e){if(r.capture||Jr.call(t,e),!e.cancelBubble)return Jn(()=>n?.call(this,e))}return e.startsWith(`pointer`)||e.startsWith(`touch`)||e===`wheel`?Ot(()=>{t.addEventListener(e,i,r)}):t.addEventListener(e,i,r),i}function Gr(e,t,n,r,i){var a={capture:r,passive:i},o=Wr(e,t,n,a);(t===document.body||t===window||t===document||t instanceof HTMLMediaElement)&&$n(()=>{t.removeEventListener(e,o,a)})}function Z(e,t,n){(t[Vr]??={})[e]=n}function Kr(e){for(var t=0;t<e.length;t++)Hr.add(e[t]);for(var n of Ur)n(e)}var qr=null;function Jr(e){var t=this,n=t.ownerDocument,r=e.type,i=e.composedPath?.()||[],a=i[0]||e.target;qr=e;var o=0,s=qr===e&&e[Vr];if(s){var c=i.indexOf(s);if(c!==-1&&(t===document||t===window)){e[Vr]=t;return}var l=i.indexOf(t);if(l===-1)return;c<=l&&(o=c)}if(a=i[o]||e.target,a!==t){ge(e,`currentTarget`,{configurable:!0,get(){return a||n}});var u=H,d=G;W(null),Sr(null);try{for(var f,p=[];a!==null;){var m=a.assignedSlot||a.parentNode||a.host||null;try{var h=a[Vr]?.[r];h!=null&&(!a.disabled||e.target===a)&&h.call(a,e)}catch(e){f?p.push(e):f=e}if(e.cancelBubble||m===t||m===null)break;a=m}if(f){for(let e of p)queueMicrotask(()=>{throw e});throw f}}finally{e[Vr]=t,delete e.currentTarget,W(u),Sr(d)}}}var Yr=globalThis?.window?.trustedTypes&&globalThis.window.trustedTypes.createPolicy(`svelte-trusted-html`,{createHTML:e=>e});function Xr(e){return Yr?.createHTML(e)??e}function Zr(e){var t=Kn(`template`);return t.innerHTML=Xr(e.replaceAll(`<!>`,`<!---->`)),t.content}function Qr(e,t){var n=G;n.nodes===null&&(n.nodes={start:e,end:t,a:null,t:null})}function Q(e,t){var n=(t&1)!=0,r=(t&2)!=0,i,a=!e.startsWith(`<!>`);return()=>{if(D)return Qr(O,null),O;i===void 0&&(i=Zr(a?e:`<!>`+e),n||(i=Vn(i)));var t=r||In?document.importNode(i,!0):i.cloneNode(!0);if(n){var o=Vn(t),s=t.lastChild;Qr(o,s)}else Qr(t,t);return t}}function $(e,t){if(D){var n=G;(!(n.f&32768)||n.nodes.end===null)&&(n.nodes.end=O),lt();return}e!==null&&e.before(t)}function $r(e,t){var n=t==null?``:typeof t==`object`?`${t}`:t;n!==(e.__t??=e.nodeValue)&&(e.__t=n,e.nodeValue=`${n}`)}function ei(e,t){return ni(e,t)}var ti=new Map;function ni(e,{target:t,anchor:n,props:r={},events:i,context:a,intro:o=!0,transformError:s}){zn();var c=void 0,l=nr(()=>{var o=n??t.appendChild(Bn());un(o,{pending:()=>{}},t=>{xt({});var n=j;if(a&&(n.c=a),i&&(r.$$events=i),D&&Qr(t,null),c=e(t,r)||{},D&&(G.nodes.end=O,O===null||O.nodeType!==8||O.data!==`]`))throw ot(),it;St()},s);var l=new Set,u=e=>{for(var n=0;n<e.length;n++){var r=e[n];if(!l.has(r)){l.add(r);var i=Br(r);for(let e of[t,document]){var a=ti.get(e);a===void 0&&(a=new Map,ti.set(e,a));var o=a.get(r);o===void 0?(e.addEventListener(r,Jr,{passive:i}),a.set(r,1)):a.set(r,o+1)}}}};return u(he(Hr)),Ur.add(u),()=>{for(var e of l)for(let n of[t,document]){var r=ti.get(n),i=r.get(e);--i==0?(n.removeEventListener(e,Jr),r.delete(e),r.size===0&&ti.delete(n)):r.set(e,i)}Ur.delete(u),o!==n&&o.parentNode?.removeChild(o)}});return ri.set(c,l),c}var ri=new WeakMap;function ii(e,t){let n=ri.get(e);return n?(ri.delete(e),n(t)):Promise.resolve()}var ai=class{anchor;#e=new Map;#t=new Map;#n=new Map;#r=new Set;#i=!0;constructor(e,t=!0){this.anchor=e,this.#i=t}#a=e=>{if(this.#e.has(e)){var t=this.#e.get(e),n=this.#t.get(t);if(n)hr(n),this.#r.delete(t);else{var r=this.#n.get(t);r&&(this.#t.set(t,r.effect),this.#n.delete(t),r.fragment.lastChild.remove(),this.anchor.before(r.fragment),n=r.effect)}for(let[t,n]of this.#e){if(this.#e.delete(t),t===e)break;let r=this.#n.get(n);r&&(V(r.effect),this.#n.delete(n))}for(let[e,r]of this.#t){if(e===t||this.#r.has(e))continue;let i=()=>{if(Array.from(this.#e.values()).includes(e)){var t=document.createDocumentFragment();_r(r,t),t.append(Bn()),this.#n.set(e,{effect:r,fragment:t})}else V(r);this.#r.delete(e),this.#t.delete(e)};this.#i||!n?(this.#r.add(e),pr(r,i,!1)):i()}}};#o=e=>{this.#e.delete(e);let t=Array.from(this.#e.values());for(let[e,n]of this.#n)t.includes(e)||(V(n.effect),this.#n.delete(e))};ensure(e,t){var n=P,r=Gn();if(t&&!this.#t.has(e)&&!this.#n.has(e))if(r){var i=document.createDocumentFragment(),a=Bn();i.append(a),this.#n.set(e,{effect:B(()=>t(a)),fragment:i})}else this.#t.set(e,B(()=>t(this.anchor)));if(this.#e.set(n,e),r){for(let[t,r]of this.#t)t===e?n.unskip_effect(r):n.skip_effect(r);for(let[t,r]of this.#n)t===e?n.unskip_effect(r.effect):n.skip_effect(r.effect);n.oncommit(this.#a),n.ondiscard(this.#o)}else D&&(this.anchor=O),this.#a(n)}};function oi(e,t,n=!1){var r;D&&(r=O,lt());var i=new ai(e),a=n?Me:0;function o(e,t){if(D){var n=ft(r);if(e!==parseInt(n.substring(1))){var a=dt();k(a),i.anchor=a,ct(!1),i.ensure(e,t),ct(!0);return}}i.ensure(e,t)}sr(()=>{var e=!1;t((t,n=0)=>{e=!0,o(n,t)}),e||o(-1,null)},a)}function si(e,t){return t}function ci(e,t,n){for(var r=[],i=t.length,a,o=t.length,s=0;s<i;s++){let n=t[s];pr(n,()=>{if(a){if(a.pending.delete(n),a.done.add(n),a.pending.size===0){var t=e.outrogroups;li(e,he(a.done)),t.delete(a),t.size===0&&(e.outrogroups=null)}}else --o},!1)}if(o===0){var c=r.length===0&&n!==null;if(c){var l=n,u=l.parentNode;Wn(u),u.append(l),e.items.clear()}li(e,t,!c)}else a={pending:new Set(t),done:new Set},(e.outrogroups??=new Set).add(a)}function li(e,t,n=!0){var r;if(e.pending.size>0){r=new Set;for(let t of e.pending.values())for(let n of t)r.add(e.items.get(n).e)}for(var i=0;i<t.length;i++){var a=t[i];r?.has(a)?(a.f|=Fe,_r(a,document.createDocumentFragment())):V(t[i],n)}}var ui;function di(e,t,n,r,i,a=null){var o=e,s=new Map;if(t&4){var c=e;o=D?k(Vn(c)):c.appendChild(Bn())}D&&lt();var l=null,u=vn(()=>{var e=n();return fe(e)?e:e==null?[]:he(e)}),d,f=new Map,p=!0;function m(e){g.effect.f&16384||(g.pending.delete(e),g.fallback=l,pi(g,d,o,t,r),l!==null&&(d.length===0?l.f&33554432?(l.f^=Fe,hi(l,null,o)):hr(l):pr(l,()=>{l=null})))}function h(e){g.pending.delete(e)}var g={effect:sr(()=>{d=X(u);var e=d.length;let c=!1;D&&ft(o)===`[!`!=(e===0)&&(o=dt(),k(o),ct(!1),c=!0);for(var g=new Set,_=P,v=Gn(),y=0;y<e;y+=1){D&&O.nodeType===8&&O.data===`]`&&(o=O,c=!0,ct(!1));var b=d[y],x=r(b,y),S=p?null:s.get(x);S?(S.v&&An(S.v,b),S.i&&An(S.i,y),v&&_.unskip_effect(S.e)):(S=mi(s,p?o:ui??=Bn(),b,x,y,i,t,n),p||(S.e.f|=Fe),s.set(x,S)),g.add(x)}if(e===0&&a&&!l&&(p?l=B(()=>a(o)):(l=B(()=>a(ui??=Bn())),l.f|=Fe)),e>g.size&&qe(``,``,``),D&&e>0&&k(dt()),!p)if(f.set(_,g),v){for(let[e,t]of s)g.has(e)||_.skip_effect(t.e);_.oncommit(m),_.ondiscard(h)}else m(_);c&&ct(!0),X(u)}),flags:t,items:s,pending:f,outrogroups:null,fallback:l};p=!1,D&&(o=O)}function fi(e){for(;e!==null&&!(e.f&32);)e=e.next;return e}function pi(e,t,n,r,i){var a=(r&8)!=0,o=t.length,s=e.items,c=fi(e.effect.first),l,u=null,d,f=[],p=[],m,h,g,_;if(a)for(_=0;_<o;_+=1)m=t[_],h=i(m,_),g=s.get(h).e,g.f&33554432||(g.nodes?.a?.measure(),(d??=new Set).add(g));for(_=0;_<o;_+=1){if(m=t[_],h=i(m,_),g=s.get(h).e,e.outrogroups!==null)for(let t of e.outrogroups)t.pending.delete(g),t.done.delete(g);if(g.f&8192&&(hr(g),a&&(g.nodes?.a?.unfix(),(d??=new Set).delete(g))),g.f&33554432)if(g.f^=Fe,g===c)hi(g,null,n);else{var v=u?u.next:c;g===e.effect.last&&(e.effect.last=g.prev),g.prev&&(g.prev.next=g.next),g.next&&(g.next.prev=g.prev),gi(e,u,g),gi(e,g,v),hi(g,v,n),u=g,f=[],p=[],c=fi(u.next);continue}if(g!==c){if(l!==void 0&&l.has(g)){if(f.length<p.length){var y=p[0],b;u=y.prev;var x=f[0],S=f[f.length-1];for(b=0;b<f.length;b+=1)hi(f[b],y,n);for(b=0;b<p.length;b+=1)l.delete(p[b]);gi(e,x.prev,S.next),gi(e,u,x),gi(e,S,y),c=y,u=S,--_,f=[],p=[]}else l.delete(g),hi(g,c,n),gi(e,g.prev,g.next),gi(e,g,u===null?e.effect.first:u.next),gi(e,u,g),u=g;continue}for(f=[],p=[];c!==null&&c!==g;)(l??=new Set).add(c),p.push(c),c=fi(c.next);if(c===null)continue}g.f&33554432||f.push(g),u=g,c=fi(g.next)}if(e.outrogroups!==null){for(let t of e.outrogroups)t.pending.size===0&&(li(e,he(t.done)),e.outrogroups?.delete(t));e.outrogroups.size===0&&(e.outrogroups=null)}if(c!==null||l!==void 0){var C=[];if(l!==void 0)for(g of l)g.f&8192||C.push(g);for(;c!==null;)!(c.f&8192)&&c!==e.fallback&&C.push(c),c=fi(c.next);var ee=C.length;if(ee>0){var te=r&4&&o===0?n:null;if(a){for(_=0;_<ee;_+=1)C[_].nodes?.a?.measure();for(_=0;_<ee;_+=1)C[_].nodes?.a?.fix()}ci(e,C,te)}}a&&Ot(()=>{if(d!==void 0)for(g of d)g.nodes?.a?.apply()})}function mi(e,t,n,r,i,a,o,s){var c=o&1?o&16?On(n):kn(n,!1,!1):null,l=o&2?On(i):null;return{v:c,i:l,e:B(()=>(a(t,c??n,l??i,s),()=>{e.delete(r)}))}}function hi(e,t,n){if(e.nodes)for(var r=e.nodes.start,i=e.nodes.end,a=t&&!(t.f&33554432)?t.nodes.start:n;r!==null;){var o=Hn(r);if(a.before(r),r===i)return;r=o}}function gi(e,t,n){t===null?e.effect.first=n:t.next=n,n===null?e.effect.last=t:n.prev=t}function _i(e,t){rr(()=>{var n=e.getRootNode(),r=n.host?n:n.head??n.ownerDocument.head;if(!r.querySelector(`#`+t.hash)){let e=Kn(`style`);e.id=t.hash,e.textContent=t.code,r.appendChild(e)}})}var vi=[...` 	
+\r\f\xA0\v﻿`];function yi(e,t,n){var r=e==null?``:``+e;if(t&&(r=r?r+` `+t:t),n){for(var i of Object.keys(n))if(n[i])r=r?r+` `+i:i;else if(r.length)for(var a=i.length,o=0;(o=r.indexOf(i,o))>=0;){var s=o+a;(o===0||vi.includes(r[o-1]))&&(s===r.length||vi.includes(r[s]))?r=(o===0?``:r.substring(0,o))+r.substring(s+1):o=s}}return r===``?null:r}function bi(e,t,n,r,i,a){var o=e.__className;if(D||o!==n||o===void 0){var s=yi(n,r,a);(!D||s!==e.getAttribute(`class`))&&(s==null?e.removeAttribute(`class`):t?e.className=s:e.setAttribute(`class`,s)),e.__className=n}else if(a&&i!==a)for(var c in a){var l=!!a[c];(i==null||l!==!!i[c])&&e.classList.toggle(c,l)}return a}var xi=Symbol(`is custom element`),Si=Symbol(`is html`),Ci=We?`link`:`LINK`;function wi(e,t,n,r){var i=Ti(e);D&&(i[t]=e.getAttribute(t),t===`src`||t===`srcset`||t===`href`&&e.nodeName===Ci)||i[t]!==(i[t]=n)&&(t===`loading`&&(e[He]=n),n==null?e.removeAttribute(t):typeof n!=`string`&&Di(e).includes(t)?e[t]=n:e.setAttribute(t,n))}function Ti(e){return e.__attributes??={[xi]:e.nodeName.includes(`-`),[Si]:e.namespaceURI===at}}var Ei=new Map;function Di(e){var t=e.getAttribute(`is`)||e.nodeName,n=Ei.get(t);if(n)return n;Ei.set(t,n=[]);for(var r,i=e,a=Element.prototype;a!==i;){for(var o in r=ve(i),r)r[o].set&&n.push(o);i=xe(i)}return n}var Oi={get(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(Ce(r)&&(r=r()),typeof r==`object`&&r&&t in r)return r[t]}},set(e,t,n){let r=e.props.length;for(;r--;){let i=e.props[r];Ce(i)&&(i=i());let a=_e(i,t);if(a&&a.set)return a.set(n),!0}return!1},getOwnPropertyDescriptor(e,t){let n=e.props.length;for(;n--;){let r=e.props[n];if(Ce(r)&&(r=r()),typeof r==`object`&&r&&t in r){let e=_e(r,t);return e&&!e.configurable&&(e.configurable=!0),e}}},has(e,t){if(t===Be||t===Ve)return!1;for(let n of e.props)if(Ce(n)&&(n=n()),n!=null&&t in n)return!0;return!1},ownKeys(e){let t=[];for(let n of e.props)if(Ce(n)&&(n=n()),n){for(let e in n)t.includes(e)||t.push(e);for(let e of Object.getOwnPropertySymbols(n))t.includes(e)||t.push(e)}return t}};function ki(...e){return new Proxy({props:e},Oi)}function Ai(e,t,n,r){var i=!_t||(n&2)!=0,a=(n&8)!=0,o=(n&16)!=0,s=r,c=!0,l=()=>(c&&(c=!1,s=o?Rr(r):r),s);let u;if(a){var d=Be in e||Ve in e;u=_e(e,t)?.set??(d&&t in e?n=>e[t]=n:void 0)}var f,p=!1;a?[f,p]=Wt(()=>e[t]):f=e[t],f===void 0&&r!==void 0&&(f=l(),u&&(i&&Qe(t),u(f)));var m=i?()=>{var n=e[t];return n===void 0?l():(c=!0,n)}:()=>{var n=e[t];return n!==void 0&&(s=void 0),n===void 0?s:n};if(i&&!(n&4))return m;if(u){var h=e.$$legacy;return(function(e,t){return arguments.length>0?((!i||!t||h||p)&&u(t?m():e),e):m()})}var g=!1,_=(n&1?gn:vn)(()=>(g=!1,m()));a&&X(_);var v=G;return(function(e,t){if(arguments.length>0){let n=t?X(_):i&&a?Pn(e):e;return R(_,n),g=!0,s!==void 0&&(s=n),e}return br&&g||v.f&16384?_.v:X(_)})}typeof window<`u`&&((window.__svelte??={}).v??=new Set).add(`5`);function ji(t){return e.Platform.isMacOS?t.metaKey:t.ctrlKey}var Mi=class{store;plugin;constructor(e,t){this.plugin=t,this.store=Lt(0),t.app.workspace.onLayoutReady(()=>{let{vault:n,metadataCache:r,workspace:i}=t.app;e.registerEvent(n.on(`create`,this.bump,this)),e.registerEvent(n.on(`delete`,this.bump,this)),e.registerEvent(n.on(`rename`,this.onRename,this)),e.registerEvent(r.on(`changed`,this.bump,this)),e.registerEvent(i.on(`periodic-notes:resolve`,this.bumpUnconditionally,this)),e.registerEvent(i.on(`periodic-notes:settings-updated`,this.bumpUnconditionally,this)),this.bump()})}bump(e){e&&!this.plugin.isPeriodic(e.path)||this.store.update(e=>e+1)}bumpUnconditionally(){this.store.update(e=>e+1)}onRename(e,t){(this.plugin.isPeriodic(e.path)||this.plugin.isPeriodic(t))&&this.store.update(e=>e+1)}getFile(e,t){return this.plugin.getPeriodicNote(t,e)}isGranularityEnabled(e){return this.plugin.settings.granularities[e]?.enabled??e===`day`}getEnabledGranularities(){return[`week`,`month`,`year`].filter(e=>this.plugin.settings.granularities[e]?.enabled)}};function Ni(e,t){return`${e}:${t.format(r[e])}`}function Pi(e,t,n){let r=new Map,i=e[1].days[0];for(let i of e){for(let e of i.days)r.set(Ni(`day`,e),t(e,`day`));if(n.includes(`week`)){let e=i.days[0];r.set(Ni(`week`,e),t(e,`week`))}}return n.includes(`month`)&&r.set(Ni(`month`,i),t(i,`month`)),n.includes(`year`)&&r.set(Ni(`year`,i),t(i,`year`)),r}var Fi=Q(`<td><div role="button" tabindex="0"> </div></td>`),Ii={hash:`svelte-1l5s8qw`,code:`.day.svelte-1l5s8qw {background-color:var(--color-background-day);border-radius:4px;color:var(--color-text-day);cursor:pointer;font-size:0.8em;height:100%;padding:4px;position:relative;text-align:center;transition:background-color 0.1s ease-in,
       color 0.1s ease-in;vertical-align:baseline;}.day.svelte-1l5s8qw:hover {background-color:var(--interactive-hover);}.day.active.svelte-1l5s8qw:hover {background-color:var(--interactive-accent-hover);}.adjacent-month.svelte-1l5s8qw {opacity:0.25;}.has-note.svelte-1l5s8qw::after {background-color:var(--text-muted);border-radius:50%;content:"";display:block;height:3px;margin:1px auto 0;width:3px;}.has-note.active.svelte-1l5s8qw::after {background-color:var(--text-on-accent);}.today.svelte-1l5s8qw {color:var(--interactive-accent);font-weight:600;}.day.svelte-1l5s8qw:active,
   .active.svelte-1l5s8qw,
-  .active.today.svelte-1l5s8qw {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function Ni(e,t){_t(t,!0),pi(e,Mi);let n=()=>Rt(o,`$displayedMonth`,r),[r,i]=zt(),a=Ei(t,`activeFilePath`,3,null),o=ht(s),c=I(()=>t.fileMap.get(ki(`day`,t.date))??null);function l(e){t.onClick?.(`day`,t.date,X(c),Di(e))}function u(e){e.target&&t.onHover?.(`day`,t.date,X(c),e.target,Di(e))}function d(e){t.onContextMenu?.(`day`,t.date,X(c),e)}var f=ji(),p=z(f);let m;var h=z(p,!0);A(p),A(f),nr((e,t)=>{m=gi(p,1,`day svelte-1l5s8qw`,null,m,e),Yr(h,t)},[()=>({active:X(c)!==null&&X(c).path===a(),"adjacent-month":!t.date.isSame(n(),`month`),"has-note":X(c)!==null,today:t.date.isSame(t.today,`day`)}),()=>t.date.format(`D`)]),Z(`click`,p,l),Z(`keydown`,p,e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),t.onClick?.(`day`,t.date,X(c),!1))}),Z(`contextmenu`,p,d),Vr(`pointerenter`,p,u),$(e,f),vt(),i()}Hr([`click`,`keydown`,`contextmenu`]);var Pi=Q(`<button type="button"><svg focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" class="svelte-he81r3"><path fill="currentColor" d="M34.52 239.03L228.87 44.69c9.37-9.37 24.57-9.37 33.94 0l22.67 22.67c9.36 9.36 9.37 24.52.04 33.9L131.49 256l154.02 154.75c9.34 9.38 9.32 24.54-.04 33.9l-22.67 22.67c-9.37 9.37-24.57 9.37-33.94 0L34.52 272.97c-9.37-9.37-9.37-24.57 0-33.94z"></path></svg></button>`),Fi={hash:`svelte-he81r3`,code:`.arrow.svelte-he81r3 {align-items:center;appearance:none;background:none;border:none;cursor:pointer;display:flex;justify-content:center;padding:0;width:24px;}.arrow.is-mobile.svelte-he81r3 {width:32px;}.right.svelte-he81r3 {transform:rotate(180deg);}.arrow.svelte-he81r3 svg:where(.svelte-he81r3) {color:var(--color-arrow);height:16px;width:16px;}`};function Ii(t,n){_t(n,!0),pi(t,Fi);var r=Pi();let i;nr(()=>{i=gi(r,1,`arrow svelte-he81r3`,null,i,{"is-mobile":e.Platform.isMobile,right:n.direction===`right`}),bi(r,`aria-label`,n.tooltip)}),Z(`click`,r,function(...e){n.onClick?.apply(this,e)}),$(t,r),vt()}Hr([`click`]);var Li=Q(`<div><span class="title svelte-16hzpd8"><span> </span> <span> </span></span></div>`),Ri={hash:`svelte-16hzpd8`,code:`.title.svelte-16hzpd8 {color:var(--color-text-title);display:flex;font-size:1.4em;gap:0.3em;margin:0;}.month.svelte-16hzpd8 {font-weight:500;}.year.svelte-16hzpd8 {color:var(--interactive-accent);}.clickable.svelte-16hzpd8 {cursor:pointer;}`};function zi(e,t){_t(t,!0),pi(e,Ri);let n=()=>Rt(a,`$displayedMonth`,r),[r,i]=zt(),a=ht(s),o=I(()=>ki(`month`,n())),c=I(()=>ki(`year`,n())),l=I(()=>t.fileMap.has(X(o))),u=I(()=>t.fileMap.has(X(c))),d=I(()=>t.fileMap.get(X(o))??null),f=I(()=>t.fileMap.get(X(c))??null);function p(e,r,i){return{click:a=>{r()?t.onClick?.(e,n(),i(),Di(a)):e===`month`&&t.resetDisplayedMonth()},hover:a=>{!r()||!a.target||t.onHover?.(e,n(),i(),a.target,Di(a))},context:a=>{let o=i();r()&&o&&t.onContextMenu?.(e,n(),o,a)}}}let m=p(`month`,()=>X(l),()=>X(d)),h=p(`year`,()=>X(u),()=>X(f));var g=Li(),_=z(g),v=z(_);let y;var b=z(v,!0);A(v);var x=zn(v,2);let S;var C=z(x,!0);A(x),A(_),A(g),nr((e,t)=>{y=gi(v,1,`month svelte-16hzpd8`,null,y,{clickable:X(l)}),bi(v,`role`,X(l)?`button`:void 0),bi(v,`tabindex`,X(l)?0:void 0),Yr(b,e),S=gi(x,1,`year svelte-16hzpd8`,null,S,{clickable:X(u)}),bi(x,`role`,X(u)?`button`:void 0),bi(x,`tabindex`,X(u)?0:void 0),Yr(C,t)},[()=>n().format(`MMM`),()=>n().format(`YYYY`)]),Z(`click`,v,function(...e){m.click?.apply(this,e)}),Z(`keydown`,v,e=>{(e.key===`Enter`||e.key===` `)&&(X(l)?t.onClick?.(`month`,n(),X(d),!1):t.resetDisplayedMonth())}),Z(`contextmenu`,v,function(...e){m.context?.apply(this,e)}),Vr(`pointerenter`,v,function(...e){m.hover?.apply(this,e)}),Z(`click`,x,function(...e){h.click?.apply(this,e)}),Z(`keydown`,x,e=>{(e.key===`Enter`||e.key===` `)&&X(u)&&t.onClick?.(`year`,n(),X(f),!1)}),Z(`contextmenu`,x,function(...e){h.context?.apply(this,e)}),Vr(`pointerenter`,x,function(...e){h.hover?.apply(this,e)}),$(e,g),vt(),i()}Hr([`click`,`keydown`,`contextmenu`]);var Bi=Q(`<div class="nav svelte-17quior"><!> <div class="right-nav svelte-17quior"><!> <button type="button">&#x25CF;</button> <!></div></div>`),Vi={hash:`svelte-17quior`,code:`.nav.svelte-17quior {align-items:baseline;display:flex;margin:0.6em 0 1em;padding:0 8px;width:100%;}.right-nav.svelte-17quior {align-items:center;display:flex;justify-content:center;margin-left:auto;}.reset-button.svelte-17quior {align-items:center;appearance:none;background:none;border:none;color:var(--color-arrow);cursor:default;display:flex;font-size:8px;opacity:0.4;padding:0.5em;}.reset-button.active.svelte-17quior {cursor:pointer;opacity:1;}`};function Hi(e,t){_t(t,!0),pi(e,Vi);let n=()=>Rt(a,`$displayedMonth`,r),[r,i]=zt(),a=ht(s);function o(){a.update(e=>e.clone().add(1,`month`))}function c(){a.update(e=>e.clone().subtract(1,`month`))}function l(){a.set(t.today.clone())}let u=I(()=>n().isSame(t.today,`month`));var d=Bi(),f=z(d);zi(f,Ti({get fileMap(){return t.fileMap},resetDisplayedMonth:l},()=>t.eventHandlers));var p=zn(f,2),m=z(p);Ii(m,{direction:`left`,onClick:c,tooltip:`Previous Month`});var h=zn(m,2);let g;Ii(zn(h,2),{direction:`right`,onClick:o,tooltip:`Next Month`}),A(p),A(d),nr(()=>{bi(h,`aria-label`,X(u)?`Current month`:`Reset to current month`),g=gi(h,1,`reset-button svelte-17quior`,null,g,{active:!X(u)})}),Z(`click`,h,l),$(e,d),vt(),i()}Hr([`click`]);function Ui(){return window.moment.weekdaysShort(!0)}function Wi(e){return e.isoWeekday()===6||e.isoWeekday()===7}function Gi(e){return e[0].clone()}function Ki(e){let t=[],n,r=e.clone().date(1),i=r.weekday(),a=r.clone().subtract(i,`days`);for(let e=0;e<42;e++)e%7==0&&(n={days:[],weekNum:a.week()},t.push(n)),n.days.push(a),a=a.clone().add(1,`days`);return t}var qi=Q(`<td class="svelte-uujyui"><div role="button" tabindex="0"> </div></td>`),Ji={hash:`svelte-uujyui`,code:`td.svelte-uujyui {border-right:1px solid var(--background-modifier-border);}.week-num.svelte-uujyui {background-color:var(--color-background-weeknum);border-radius:4px;color:var(--color-text-weeknum);cursor:pointer;font-size:0.65em;height:100%;padding:4px;text-align:center;transition:background-color 0.1s ease-in,
-      color 0.1s ease-in;vertical-align:baseline;}.week-num.svelte-uujyui:hover {background-color:var(--interactive-hover);}.week-num.active.svelte-uujyui:hover {background-color:var(--interactive-accent-hover);}.active.svelte-uujyui {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function Yi(e,t){_t(t,!0),pi(e,Ji);let n=Ei(t,`activeFilePath`,3,null),r=I(()=>Gi(t.days)),i=I(()=>t.fileMap.get(ki(`week`,X(r)))??null);function a(e){e.target&&t.onHover?.(`week`,X(r),X(i),e.target,Di(e))}var o=qi(),s=z(o);let c;var l=z(s,!0);A(s),A(o),nr(()=>{c=gi(s,1,`week-num svelte-uujyui`,null,c,{active:X(i)!==null&&X(i).path===n()}),Yr(l,t.weekNum)}),Z(`click`,s,function(...e){(t.onClick&&(e=>t.onClick(`week`,X(r),X(i),Di(e))))?.apply(this,e)}),Z(`keydown`,s,function(...e){(t.onClick&&(e=>(e.key===`Enter`||e.key===` `)&&t.onClick(`week`,X(r),X(i),!1)))?.apply(this,e)}),Z(`contextmenu`,s,function(...e){(t.onContextMenu&&(e=>t.onContextMenu(`week`,X(r),X(i),e)))?.apply(this,e)}),Vr(`pointerenter`,s,a),$(e,o),vt()}Hr([`click`,`keydown`,`contextmenu`]);var Xi=Q(`<col/>`),Zi=Q(`<col/>`),Qi=Q(`<th class="svelte-5j20fs">W</th>`),$i=Q(`<th class="svelte-5j20fs"> </th>`),ea=Q(`<tr><!><!></tr>`),ta=Q(`<div id="calendar-container" class="container svelte-5j20fs"><!> <table class="calendar svelte-5j20fs"><colgroup><!><!></colgroup><thead><tr><!><!></tr></thead><tbody></tbody></table></div>`),na={hash:`svelte-5j20fs`,code:`.container.svelte-5j20fs {--color-background-heading: transparent;--color-background-day: transparent;--color-background-weeknum: transparent;--color-background-weekend: transparent;--color-arrow: var(--text-muted);--color-button: var(--text-muted);--color-text-title: var(--text-normal);--color-text-heading: var(--text-muted);--color-text-day: var(--text-normal);--color-text-today: var(--interactive-accent);--color-text-weeknum: var(--text-muted);}.container.svelte-5j20fs {padding:0 8px;}.weekend.svelte-5j20fs {background-color:var(--color-background-weekend);}.calendar.svelte-5j20fs {border-collapse:collapse;width:100%;}th.svelte-5j20fs {background-color:var(--color-background-heading);color:var(--color-text-heading);font-size:0.6em;letter-spacing:1px;padding:4px;text-align:center;text-transform:uppercase;}`};function ra(e,t){_t(t,!0),pi(e,na);let n=()=>Rt(c,`$displayedMonthStore`,r),[r,i]=zt(),a=L(null),o=L(window.moment()),c=Nt(window.moment());gt(s,c);let l=L(Ki(window.moment())),u=L(!1),d=L(new Map);Xn(()=>{R(l,Ki(n()))}),Xn(()=>{let e=X(l);return t.fileStore.store.subscribe(()=>{R(u,t.fileStore.isGranularityEnabled(`week`),!0),R(d,Ai(e,(e,n)=>t.fileStore.getFile(e,n),t.fileStore.getEnabledGranularities()))})});let f=I(()=>({onHover:t.onHover,onClick:t.onClick,onContextMenu:t.onContextMenu})),p=Ui();function m(){let e=window.moment();e.isSame(X(o),`day`)||R(o,e)}function h(e){R(a,e,!0)}var g={tick:m,setActiveFilePath:h},_=ta(),v=z(_);Hi(v,{get fileMap(){return X(d)},get today(){return X(o)},get eventHandlers(){return X(f)}});var y=zn(v,2),b=z(y),x=z(b),S=e=>{$(e,Xi())};ni(x,e=>{X(u)&&e(S)}),si(zn(x),17,()=>X(l)[1].days,ri,(e,t)=>{var n=Zi();let r;nr(e=>r=gi(n,1,`svelte-5j20fs`,null,r,e),[()=>({weekend:Wi(X(t))})]),$(e,n)}),A(b);var C=zn(b),ee=z(C),te=z(ee),ne=e=>{$(e,Qi())};ni(te,e=>{X(u)&&e(ne)}),si(zn(te),17,()=>p,ri,(e,t)=>{var n=$i(),r=z(n,!0);A(n),nr(()=>Yr(r,X(t))),$(e,n)}),A(ee),A(C);var re=zn(C);si(re,21,()=>X(l),e=>ki(`week`,e.days[0]),(e,t)=>{var n=ea(),r=z(n),i=e=>{Yi(e,Ti({get fileMap(){return X(d)},get activeFilePath(){return X(a)}},()=>X(t),()=>X(f)))};ni(r,e=>{X(u)&&e(i)}),si(zn(r),17,()=>X(t).days,e=>e.format(),(e,t)=>{Ni(e,Ti({get date(){return X(t)},get fileMap(){return X(d)},get today(){return X(o)},get activeFilePath(){return X(a)}},()=>X(f)))}),A(n),$(e,n)}),A(re),A(y),A(_),$(e,_);var ie=vt(g);return i(),ie}var ia=class extends e.ItemView{calendar;plugin;constructor(e,t){super(e),this.plugin=t,this.registerEvent(this.app.workspace.on(`file-open`,this.onFileOpen.bind(this)))}getViewType(){return o}getDisplayText(){return`Calendar`}getIcon(){return`calendar-day`}async onClose(){this.calendar&&ei(this.calendar)}async onOpen(){let e=new Oi(this,this.plugin),t=Xr(ra,{target:this.contentEl,props:{fileStore:e,onHover:this.onHover.bind(this),onClick:this.onClick.bind(this),onContextMenu:this.onContextMenu.bind(this)}});if(!(`tick`in t&&`setActiveFilePath`in t))throw Error(`Calendar component missing expected exports`);this.calendar=t}onHover(e,t,n,r,i){if(!i)return;let a=t.format(e===`day`?`YYYY-MM-DD`:t.localeData().longDateFormat(`L`));this.app.workspace.trigger(`link-hover`,this,r,a,n?.path??``)}onClick(e,t,n,r){this.plugin.openPeriodicNote(e,t,{inNewSplit:r})}onContextMenu(t,n,r,i){if(!r)return;let a=new e.Menu;a.addItem(e=>e.setTitle(`Delete`).setIcon(`trash`).onClick(()=>{this.app.vault.trash(r,!0)})),this.app.workspace.trigger(`file-menu`,a,r,`calendar-context-menu`,null),a.showAtPosition({x:i.pageX,y:i.pageY})}onFileOpen(e){if(this.app.workspace.layoutReady&&this.calendar){let e=this.app.workspace.getActiveFile()?.path??null;this.calendar.setActiveFilePath(e),this.calendar.tick()}}},aa={day:{periodicity:`daily`,relativeUnit:`today`,labelOpenPresent:`Open today's daily note`},week:{periodicity:`weekly`,relativeUnit:`this week`,labelOpenPresent:`Open this week's note`},month:{periodicity:`monthly`,relativeUnit:`this month`,labelOpenPresent:`Open this month's note`},year:{periodicity:`yearly`,relativeUnit:`this year`,labelOpenPresent:`Open this year's note`}};async function oa(t,n,r){let i=t.workspace.getActiveFile();if(!i)return;let a=n.findInCache(i.path);if(!a)return;let o=n.findAdjacent(i.path,r);if(o){let n=t.vault.getAbstractFileByPath(o.filePath);n&&n instanceof e.TFile&&await t.workspace.getLeaf().openFile(n,{active:!0})}else{let t=r===`forwards`?`after`:`before`;new e.Notice(`There's no ${aa[a.granularity].periodicity} note ${t} this`)}}async function sa(e,t,n){let r=e.workspace.getActiveFile();if(!r)return;let i=t.findInCache(r.path);if(!i)return;let a=n===`forwards`?1:-1,o=i.date.clone().add(a,i.granularity);t.openPeriodicNote(i.granularity,o)}function ca(e,t,n){let r=aa[n];return[{id:`open-${r.periodicity}-note`,name:r.labelOpenPresent,checkCallback:e=>{if(!t.settings.granularities[n].enabled)return!1;if(e)return!0;t.openPeriodicNote(n,window.moment())}},{id:`next-${r.periodicity}-note`,name:`Jump forwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;oa(e,t,`forwards`)}},{id:`prev-${r.periodicity}-note`,name:`Jump backwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;oa(e,t,`backwards`)}},{id:`open-next-${r.periodicity}-note`,name:`Open next ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;sa(e,t,`forwards`)}},{id:`open-prev-${r.periodicity}-note`,name:`Open previous ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;sa(e,t,`backwards`)}}]}function la(t,n){let r=new e.Menu,i=c.filter(e=>t.settings.granularities[e].enabled);for(let e of i){let n=aa[e];r.addItem(r=>r.setTitle(n.labelOpenPresent).setIcon(`calendar-${e}`).onClick(()=>t.openPeriodicNote(e,window.moment())))}r.showAtPosition(n)}var ua=`
+  .active.today.svelte-1l5s8qw {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function Li(e,t){xt(t,!0),_i(e,Ii);let n=()=>Ht(o,`$displayedMonth`,r),[r,i]=Ut(),a=Ai(t,`activeFilePath`,3,null),o=yt(c),s=I(()=>t.fileMap.get(Ni(`day`,t.date))??null);function l(e){t.onClick?.(`day`,t.date,X(s),ji(e))}function u(e){e.target&&t.onHover?.(`day`,t.date,X(s),e.target,ji(e))}function d(e){t.onContextMenu?.(`day`,t.date,X(s),e)}var f=Fi(),p=z(f);let m;var h=z(p,!0);A(p),A(f),or((e,t)=>{m=bi(p,1,`day svelte-1l5s8qw`,null,m,e),$r(h,t)},[()=>({active:X(s)!==null&&X(s).path===a(),"adjacent-month":!t.date.isSame(n(),`month`),"has-note":X(s)!==null,today:t.date.isSame(t.today,`day`)}),()=>t.date.format(`D`)]),Z(`click`,p,l),Z(`keydown`,p,e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),t.onClick?.(`day`,t.date,X(s),!1))}),Z(`contextmenu`,p,d),Gr(`pointerenter`,p,u),$(e,f),St(),i()}Kr([`click`,`keydown`,`contextmenu`]);var Ri=Q(`<button type="button"><svg focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" class="svelte-he81r3"><path fill="currentColor" d="M34.52 239.03L228.87 44.69c9.37-9.37 24.57-9.37 33.94 0l22.67 22.67c9.36 9.36 9.37 24.52.04 33.9L131.49 256l154.02 154.75c9.34 9.38 9.32 24.54-.04 33.9l-22.67 22.67c-9.37 9.37-24.57 9.37-33.94 0L34.52 272.97c-9.37-9.37-9.37-24.57 0-33.94z"></path></svg></button>`),zi={hash:`svelte-he81r3`,code:`.arrow.svelte-he81r3 {align-items:center;appearance:none;background:none;border:none;cursor:pointer;display:flex;justify-content:center;padding:0;width:24px;}.arrow.is-mobile.svelte-he81r3 {width:32px;}.right.svelte-he81r3 {transform:rotate(180deg);}.arrow.svelte-he81r3 svg:where(.svelte-he81r3) {color:var(--color-arrow);height:16px;width:16px;}`};function Bi(t,n){xt(n,!0),_i(t,zi);var r=Ri();let i;or(()=>{i=bi(r,1,`arrow svelte-he81r3`,null,i,{"is-mobile":e.Platform.isMobile,right:n.direction===`right`}),wi(r,`aria-label`,n.tooltip)}),Z(`click`,r,function(...e){n.onClick?.apply(this,e)}),$(t,r),St()}Kr([`click`]);var Vi=Q(`<div><span class="title svelte-16hzpd8"><span> </span> <span> </span></span></div>`),Hi={hash:`svelte-16hzpd8`,code:`.title.svelte-16hzpd8 {color:var(--color-text-title);display:flex;font-size:1.4em;gap:0.3em;margin:0;}.month.svelte-16hzpd8 {font-weight:500;}.year.svelte-16hzpd8 {color:var(--interactive-accent);}.clickable.svelte-16hzpd8 {cursor:pointer;}`};function Ui(e,t){xt(t,!0),_i(e,Hi);let n=()=>Ht(a,`$displayedMonth`,r),[r,i]=Ut(),a=yt(c),o=I(()=>Ni(`month`,n())),s=I(()=>Ni(`year`,n())),l=I(()=>t.fileMap.has(X(o))),u=I(()=>t.fileMap.has(X(s))),d=I(()=>t.fileMap.get(X(o))??null),f=I(()=>t.fileMap.get(X(s))??null);function p(e,r,i){return{click:a=>{r()?t.onClick?.(e,n(),i(),ji(a)):e===`month`&&t.resetDisplayedMonth()},hover:a=>{!r()||!a.target||t.onHover?.(e,n(),i(),a.target,ji(a))},context:a=>{let o=i();r()&&o&&t.onContextMenu?.(e,n(),o,a)}}}let m=p(`month`,()=>X(l),()=>X(d)),h=p(`year`,()=>X(u),()=>X(f));var g=Vi(),_=z(g),v=z(_);let y;var b=z(v,!0);A(v);var x=Un(v,2);let S;var C=z(x,!0);A(x),A(_),A(g),or((e,t)=>{y=bi(v,1,`month svelte-16hzpd8`,null,y,{clickable:X(l)}),wi(v,`role`,X(l)?`button`:void 0),wi(v,`tabindex`,X(l)?0:void 0),$r(b,e),S=bi(x,1,`year svelte-16hzpd8`,null,S,{clickable:X(u)}),wi(x,`role`,X(u)?`button`:void 0),wi(x,`tabindex`,X(u)?0:void 0),$r(C,t)},[()=>n().format(`MMM`),()=>n().format(`YYYY`)]),Z(`click`,v,function(...e){m.click?.apply(this,e)}),Z(`keydown`,v,e=>{(e.key===`Enter`||e.key===` `)&&(X(l)?t.onClick?.(`month`,n(),X(d),!1):t.resetDisplayedMonth())}),Z(`contextmenu`,v,function(...e){m.context?.apply(this,e)}),Gr(`pointerenter`,v,function(...e){m.hover?.apply(this,e)}),Z(`click`,x,function(...e){h.click?.apply(this,e)}),Z(`keydown`,x,e=>{(e.key===`Enter`||e.key===` `)&&X(u)&&t.onClick?.(`year`,n(),X(f),!1)}),Z(`contextmenu`,x,function(...e){h.context?.apply(this,e)}),Gr(`pointerenter`,x,function(...e){h.hover?.apply(this,e)}),$(e,g),St(),i()}Kr([`click`,`keydown`,`contextmenu`]);var Wi=Q(`<div class="nav svelte-17quior"><!> <div class="right-nav svelte-17quior"><!> <button type="button">&#x25CF;</button> <!></div></div>`),Gi={hash:`svelte-17quior`,code:`.nav.svelte-17quior {align-items:baseline;display:flex;margin:0.6em 0 1em;padding:0 8px;width:100%;}.right-nav.svelte-17quior {align-items:center;display:flex;justify-content:center;margin-left:auto;}.reset-button.svelte-17quior {align-items:center;appearance:none;background:none;border:none;color:var(--color-arrow);cursor:default;display:flex;font-size:8px;opacity:0.4;padding:0.5em;}.reset-button.active.svelte-17quior {cursor:pointer;opacity:1;}`};function Ki(e,t){xt(t,!0),_i(e,Gi);let n=()=>Ht(a,`$displayedMonth`,r),[r,i]=Ut(),a=yt(c);function o(){a.update(e=>e.clone().add(1,`month`))}function s(){a.update(e=>e.clone().subtract(1,`month`))}function l(){a.set(t.today.clone())}let u=I(()=>n().isSame(t.today,`month`));var d=Wi(),f=z(d);Ui(f,ki({get fileMap(){return t.fileMap},resetDisplayedMonth:l},()=>t.eventHandlers));var p=Un(f,2),m=z(p);Bi(m,{direction:`left`,onClick:s,tooltip:`Previous Month`});var h=Un(m,2);let g;Bi(Un(h,2),{direction:`right`,onClick:o,tooltip:`Next Month`}),A(p),A(d),or(()=>{wi(h,`aria-label`,X(u)?`Current month`:`Reset to current month`),g=bi(h,1,`reset-button svelte-17quior`,null,g,{active:!X(u)})}),Z(`click`,h,l),$(e,d),St(),i()}Kr([`click`]);function qi(){return window.moment.weekdaysShort(!0)}function Ji(e){return e.isoWeekday()===6||e.isoWeekday()===7}function Yi(e){return e[0].clone()}function Xi(e){let t=[],n,r=e.clone().date(1),i=r.weekday(),a=r.clone().subtract(i,`days`);for(let e=0;e<42;e++)e%7==0&&(n={days:[],weekNum:a.week()},t.push(n)),n.days.push(a),a=a.clone().add(1,`days`);return t}var Zi=Q(`<td class="svelte-uujyui"><div role="button" tabindex="0"> </div></td>`),Qi={hash:`svelte-uujyui`,code:`td.svelte-uujyui {border-right:1px solid var(--background-modifier-border);}.week-num.svelte-uujyui {background-color:var(--color-background-weeknum);border-radius:4px;color:var(--color-text-weeknum);cursor:pointer;font-size:0.65em;height:100%;padding:4px;text-align:center;transition:background-color 0.1s ease-in,
+      color 0.1s ease-in;vertical-align:baseline;}.week-num.svelte-uujyui:hover {background-color:var(--interactive-hover);}.week-num.active.svelte-uujyui:hover {background-color:var(--interactive-accent-hover);}.active.svelte-uujyui {color:var(--text-on-accent);background-color:var(--interactive-accent);}`};function $i(e,t){xt(t,!0),_i(e,Qi);let n=Ai(t,`activeFilePath`,3,null),r=I(()=>Yi(t.days)),i=I(()=>t.fileMap.get(Ni(`week`,X(r)))??null);function a(e){e.target&&t.onHover?.(`week`,X(r),X(i),e.target,ji(e))}var o=Zi(),s=z(o);let c;var l=z(s,!0);A(s),A(o),or(()=>{c=bi(s,1,`week-num svelte-uujyui`,null,c,{active:X(i)!==null&&X(i).path===n()}),$r(l,t.weekNum)}),Z(`click`,s,function(...e){(t.onClick&&(e=>t.onClick(`week`,X(r),X(i),ji(e))))?.apply(this,e)}),Z(`keydown`,s,function(...e){(t.onClick&&(e=>(e.key===`Enter`||e.key===` `)&&t.onClick(`week`,X(r),X(i),!1)))?.apply(this,e)}),Z(`contextmenu`,s,function(...e){(t.onContextMenu&&(e=>t.onContextMenu(`week`,X(r),X(i),e)))?.apply(this,e)}),Gr(`pointerenter`,s,a),$(e,o),St()}Kr([`click`,`keydown`,`contextmenu`]);var ea=Q(`<col/>`),ta=Q(`<col/>`),na=Q(`<th class="svelte-5j20fs">W</th>`),ra=Q(`<th class="svelte-5j20fs"> </th>`),ia=Q(`<tr><!><!></tr>`),aa=Q(`<div id="calendar-container" class="container svelte-5j20fs"><!> <table class="calendar svelte-5j20fs"><colgroup><!><!></colgroup><thead><tr><!><!></tr></thead><tbody></tbody></table></div>`),oa={hash:`svelte-5j20fs`,code:`.container.svelte-5j20fs {--color-background-heading: transparent;--color-background-day: transparent;--color-background-weeknum: transparent;--color-background-weekend: transparent;--color-arrow: var(--text-muted);--color-button: var(--text-muted);--color-text-title: var(--text-normal);--color-text-heading: var(--text-muted);--color-text-day: var(--text-normal);--color-text-today: var(--interactive-accent);--color-text-weeknum: var(--text-muted);}.container.svelte-5j20fs {padding:0 8px;}.weekend.svelte-5j20fs {background-color:var(--color-background-weekend);}.calendar.svelte-5j20fs {border-collapse:collapse;width:100%;}th.svelte-5j20fs {background-color:var(--color-background-heading);color:var(--color-text-heading);font-size:0.6em;letter-spacing:1px;padding:4px;text-align:center;text-transform:uppercase;}`};function sa(e,t){xt(t,!0),_i(e,oa);let n=()=>Ht(s,`$displayedMonthStore`,r),[r,i]=Ut(),a=L(null),o=L(window.moment()),s=Lt(window.moment());bt(c,s);let l=L(Xi(window.moment())),u=L(!1),d=L(new Map);er(()=>{R(l,Xi(n()))}),er(()=>{let e=X(l);return t.fileStore.store.subscribe(()=>{R(u,t.fileStore.isGranularityEnabled(`week`),!0),R(d,Pi(e,(e,n)=>t.fileStore.getFile(e,n),t.fileStore.getEnabledGranularities()))})});let f=I(()=>({onHover:t.onHover,onClick:t.onClick,onContextMenu:t.onContextMenu})),p=qi();function m(){let e=window.moment();e.isSame(X(o),`day`)||R(o,e)}function h(e){R(a,e,!0)}var g={tick:m,setActiveFilePath:h},_=aa(),v=z(_);Ki(v,{get fileMap(){return X(d)},get today(){return X(o)},get eventHandlers(){return X(f)}});var y=Un(v,2),b=z(y),x=z(b),S=e=>{$(e,ea())};oi(x,e=>{X(u)&&e(S)}),di(Un(x),17,()=>X(l)[1].days,si,(e,t)=>{var n=ta();let r;or(e=>r=bi(n,1,`svelte-5j20fs`,null,r,e),[()=>({weekend:Ji(X(t))})]),$(e,n)}),A(b);var C=Un(b),ee=z(C),te=z(ee),ne=e=>{$(e,na())};oi(te,e=>{X(u)&&e(ne)}),di(Un(te),17,()=>p,si,(e,t)=>{var n=ra(),r=z(n,!0);A(n),or(()=>$r(r,X(t))),$(e,n)}),A(ee),A(C);var re=Un(C);di(re,21,()=>X(l),e=>Ni(`week`,e.days[0]),(e,t)=>{var n=ia(),r=z(n),i=e=>{$i(e,ki({get fileMap(){return X(d)},get activeFilePath(){return X(a)}},()=>X(t),()=>X(f)))};oi(r,e=>{X(u)&&e(i)}),di(Un(r),17,()=>X(t).days,e=>e.format(),(e,t)=>{Li(e,ki({get date(){return X(t)},get fileMap(){return X(d)},get today(){return X(o)},get activeFilePath(){return X(a)}},()=>X(f)))}),A(n),$(e,n)}),A(re),A(y),A(_),$(e,_);var ie=St(g);return i(),ie}var ca=class extends e.ItemView{calendar;plugin;constructor(e,t){super(e),this.plugin=t,this.registerEvent(this.app.workspace.on(`file-open`,this.onFileOpen.bind(this)))}getViewType(){return s}getDisplayText(){return`Calendar`}getIcon(){return`calendar-day`}async onClose(){this.calendar&&ii(this.calendar)}async onOpen(){let e=new Mi(this,this.plugin),t=ei(sa,{target:this.contentEl,props:{fileStore:e,onHover:this.onHover.bind(this),onClick:this.onClick.bind(this),onContextMenu:this.onContextMenu.bind(this)}});if(!(`tick`in t&&`setActiveFilePath`in t))throw Error(`Calendar component missing expected exports`);this.calendar=t}onHover(e,t,n,r,i){if(!i)return;let a=t.format(e===`day`?`YYYY-MM-DD`:t.localeData().longDateFormat(`L`));this.app.workspace.trigger(`link-hover`,this,r,a,n?.path??``)}onClick(e,t,n,r){this.plugin.openPeriodicNote(e,t,{inNewSplit:r})}onContextMenu(t,n,r,i){if(!r)return;let a=new e.Menu;a.addItem(e=>e.setTitle(`Delete`).setIcon(`trash`).onClick(()=>{this.app.vault.trash(r,!0)})),this.app.workspace.trigger(`file-menu`,a,r,`calendar-context-menu`,null),a.showAtPosition({x:i.pageX,y:i.pageY})}onFileOpen(e){if(this.app.workspace.layoutReady&&this.calendar){let e=this.app.workspace.getActiveFile()?.path??null;this.calendar.setActiveFilePath(e),this.calendar.tick()}}},la={day:{periodicity:`daily`,relativeUnit:`today`,labelOpenPresent:`Open today's daily note`},week:{periodicity:`weekly`,relativeUnit:`this week`,labelOpenPresent:`Open this week's note`},month:{periodicity:`monthly`,relativeUnit:`this month`,labelOpenPresent:`Open this month's note`},year:{periodicity:`yearly`,relativeUnit:`this year`,labelOpenPresent:`Open this year's note`}};async function ua(t,n,r){let i=t.workspace.getActiveFile();if(!i)return;let a=n.findInCache(i.path);if(!a)return;let o=n.findAdjacent(i.path,r);if(o){let n=t.vault.getAbstractFileByPath(o.filePath);n&&n instanceof e.TFile&&await t.workspace.getLeaf().openFile(n,{active:!0})}else{let t=r===`forwards`?`after`:`before`;new e.Notice(`There's no ${la[a.granularity].periodicity} note ${t} this`)}}async function da(e,t,n){let r=e.workspace.getActiveFile();if(!r)return;let i=t.findInCache(r.path);if(!i)return;let a=n===`forwards`?1:-1,o=i.date.clone().add(a,i.granularity);t.openPeriodicNote(i.granularity,o)}function fa(e,t,n){let r=la[n];return[{id:`open-${r.periodicity}-note`,name:r.labelOpenPresent,checkCallback:e=>{if(!t.settings.granularities[n].enabled)return!1;if(e)return!0;t.openPeriodicNote(n,window.moment())}},{id:`next-${r.periodicity}-note`,name:`Jump forwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;ua(e,t,`forwards`)}},{id:`prev-${r.periodicity}-note`,name:`Jump backwards to closest ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;ua(e,t,`backwards`)}},{id:`open-next-${r.periodicity}-note`,name:`Open next ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;da(e,t,`forwards`)}},{id:`open-prev-${r.periodicity}-note`,name:`Open previous ${r.periodicity} note`,checkCallback:r=>{if(!t.settings.granularities[n].enabled)return!1;let i=e.workspace.getActiveFile();if(r)return i?t.isPeriodic(i.path,n):!1;da(e,t,`backwards`)}}]}function pa(t,n){let r=new e.Menu,i=l.filter(e=>t.settings.granularities[e].enabled);for(let e of i){let n=la[e];r.addItem(r=>r.setTitle(n.labelOpenPresent).setIcon(`calendar-${e}`).onClick(()=>t.openPeriodicNote(e,window.moment())))}r.showAtPosition(n)}var ma=`
 <g>
 <path d="M24.78 3C22.646 3 20.9 4.746 20.9 6.88V10.76H9.26C7.223 10.76 5.38 12.312 5.38 14.543V92.628C5.38 93.695 6.059 94.859 6.835 95.344C7.611 95.926 8.387 96.12 9.26 96.12H90.74C91.613 96.12 92.389 95.926 93.165 95.344C93.941 94.762 94.62 93.695 94.62 92.628V14.543C94.62 12.506 92.971 10.76 90.934 10.76H79.1V6.88C79.1 4.746 77.354 3 75.22 3H71.34C69.206 3 67.46 4.746 67.46 6.88V10.76H32.54V6.88C32.54 4.746 30.794 3 28.66 3H24.78ZM24.78 6.88H28.66V18.52H24.78V6.88ZM71.34 6.88H75.22V18.52H71.34V6.88ZM9.26 14.64H20.9V18.52C20.9 20.654 22.646 22.4 24.78 22.4H28.66C30.794 22.4 32.54 20.654 32.54 18.52V14.64H67.46V18.52C67.46 20.654 69.206 22.4 71.34 22.4H75.22C77.354 22.4 79.1 20.654 79.1 18.52V14.64H90.74V28.22H9.26V14.64ZM9.26 32.1H90.74V92.24H9.26V32.1Z" fill="currentColor" stroke-width="1" stroke="currentColor"/>
 <path d="M55.2539 79.0024H49.3613V55.3319C49.3613 52.5068 49.4282 50.2668 49.5619 48.6119C49.1775 49.0131 48.701 49.4561 48.1327 49.9408C47.581 50.4256 45.7088 51.9635 42.516 54.5546L39.5571 50.8185L50.3393 42.3432H55.2539V79.0024Z" fill="currentColor"/>
 </g>
-`,da=`
+`,ha=`
 <g>
 <path d="M24.78 3C22.646 3 20.9 4.746 20.9 6.88V10.76H9.26C7.223 10.76 5.38 12.312 5.38 14.543V92.628C5.38 93.695 6.059 94.859 6.835 95.344C7.611 95.926 8.387 96.12 9.26 96.12H90.74C91.613 96.12 92.389 95.926 93.165 95.344C93.941 94.762 94.62 93.695 94.62 92.628V14.543C94.62 12.506 92.971 10.76 90.934 10.76H79.1V6.88C79.1 4.746 77.354 3 75.22 3H71.34C69.206 3 67.46 4.746 67.46 6.88V10.76H32.54V6.88C32.54 4.746 30.794 3 28.66 3H24.78ZM24.78 6.88H28.66V18.52H24.78V6.88ZM71.34 6.88H75.22V18.52H71.34V6.88ZM9.26 14.64H20.9V18.52C20.9 20.654 22.646 22.4 24.78 22.4H28.66C30.794 22.4 32.54 20.654 32.54 18.52V14.64H67.46V18.52C67.46 20.654 69.206 22.4 71.34 22.4H75.22C77.354 22.4 79.1 20.654 79.1 18.52V14.64H90.74V28.22H9.26V14.64ZM9.26 32.1H90.74V92.24H9.26V32.1Z" fill="currentColor" stroke-width="1" stroke="currentColor"/>
 <path d="M42.8799 78.3604L56.5679 48.6873H38.5698V43.7852H62.512V47.669L48.895 78.3604H42.8799Z" fill="currentColor"/>
 </g>
-`,fa=`
+`,ga=`
 <g>
 <path d="M24.78 3C22.646 3 20.9 4.746 20.9 6.88V10.76H9.26C7.223 10.76 5.38 12.312 5.38 14.543V92.628C5.38 93.695 6.059 94.859 6.835 95.344C7.611 95.926 8.387 96.12 9.26 96.12H90.74C91.613 96.12 92.389 95.926 93.165 95.344C93.941 94.762 94.62 93.695 94.62 92.628V14.543C94.62 12.506 92.971 10.76 90.934 10.76H79.1V6.88C79.1 4.746 77.354 3 75.22 3H71.34C69.206 3 67.46 4.746 67.46 6.88V10.76H32.54V6.88C32.54 4.746 30.794 3 28.66 3H24.78ZM24.78 6.88H28.66V18.52H24.78V6.88ZM71.34 6.88H75.22V18.52H71.34V6.88ZM9.26 14.64H20.9V18.52C20.9 20.654 22.646 22.4 24.78 22.4H28.66C30.794 22.4 32.54 20.654 32.54 18.52V14.64H67.46V18.52C67.46 20.654 69.206 22.4 71.34 22.4H75.22C77.354 22.4 79.1 20.654 79.1 18.52V14.64H90.74V28.22H9.26V14.64ZM9.26 32.1H90.74V92.24H9.26V32.1Z" fill="currentColor" stroke-width="1" stroke="currentColor"/>
 <path d="M51.3075 52.8546C51.3075 54.9201 50.7057 56.6437 49.5022 58.0256C48.2986 59.3926 46.6046 60.3139 44.4204 60.7894V60.9677C47.0356 61.2946 48.9969 62.1118 50.3045 63.4194C51.6121 64.7122 52.2659 66.4358 52.2659 68.5904C52.2659 71.7257 51.1589 74.1477 48.9449 75.8565C46.7309 77.5504 43.5808 78.3974 39.4946 78.3974C35.8838 78.3974 32.8377 77.8105 30.3562 76.6366V71.9783C31.7381 72.6618 33.2018 73.1893 34.7471 73.5608C36.2924 73.9322 37.7784 74.118 39.2048 74.118C41.7309 74.118 43.618 73.6499 44.8661 72.7138C46.1143 71.7777 46.7384 70.3289 46.7384 68.3675C46.7384 66.629 46.0474 65.3511 44.6655 64.5339C43.2836 63.7166 41.1142 63.308 38.1573 63.308H35.3266V59.0509H38.2018C43.4025 59.0509 46.0028 57.2529 46.0028 53.657C46.0028 52.2603 45.5496 51.183 44.6432 50.4252C43.7368 49.6674 42.3995 49.2885 40.6313 49.2885C39.398 49.2885 38.2093 49.4668 37.0651 49.8234C35.921 50.1652 34.5688 50.8412 33.0086 51.8517L30.4454 48.1963C33.4321 45.9972 36.9017 44.8976 40.8542 44.8976C44.138 44.8976 46.7012 45.6034 48.5437 47.015C50.3863 48.4266 51.3075 50.3732 51.3075 52.8546Z" fill="currentColor"/>
 <path d="M69.6199 77.9516H64.382V56.9112C64.382 54.4 64.4415 52.4089 64.5603 50.9378C64.2186 51.2944 63.7951 51.6882 63.2899 52.1191C62.7995 52.55 61.1353 53.9171 58.2972 56.2202L55.6672 52.8992L65.2513 45.3657H69.6199V77.9516Z" fill="currentColor"/>
 </g>
-`,pa=`
+`,_a=`
 <g>
 <path d="M24.768 3C22.634 3 20.888 4.746 20.888 6.88V10.76H9.24804C7.21104 10.76 5.36804 12.312 5.36804 14.543V92.628C5.36804 93.695 6.04704 94.859 6.82304 95.344C7.59904 95.926 8.37504 96.12 9.24804 96.12H90.728C91.601 96.12 92.377 95.926 93.153 95.344C93.929 94.762 94.608 93.695 94.608 92.628V14.543C94.608 12.506 92.959 10.76 90.922 10.76H79.088V6.88C79.088 4.746 77.342 3 75.208 3H71.328C69.194 3 67.448 4.746 67.448 6.88V10.76H32.528V6.88C32.528 4.746 30.782 3 28.648 3H24.768ZM24.768 6.88H28.648V18.52H24.768V6.88ZM71.328 6.88H75.208V18.52H71.328V6.88ZM9.24804 14.64H20.888V18.52C20.888 20.654 22.634 22.4 24.768 22.4H28.648C30.782 22.4 32.528 20.654 32.528 18.52V14.64H67.448V18.52C67.448 20.654 69.194 22.4 71.328 22.4H75.208C77.342 22.4 79.088 20.654 79.088 18.52V14.64H90.728V28.22H9.24804V14.64ZM9.24804 32.1H90.728V92.24H9.24804V32.1Z" fill="currentColor" stroke="currentColor" stroke-width="0.17"/>
 <path d="M49.2303 60.2321L56.9421 45.3656H62.7371L51.8826 65.3139V77.9515H46.5333V65.4922L35.7234 45.3656H41.5184L49.2303 60.2321Z" fill="currentColor"/>
 </g>
-`,ma=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getMarkdownFiles().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}},ha=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getAllFolders().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}};function ga(e,t){return t?e.metadataCache.getFirstLinkpathDest(t,``)?``:`Template file not found`:``}function _a(t,n){return!n||n===`/`||t.vault.getAbstractFileByPath((0,e.normalizePath)(n))?``:`Folder not found in vault`}var va={day:`Daily Notes`,week:`Weekly Notes`,month:`Monthly Notes`,year:`Yearly Notes`},ya=class extends e.PluginSettingTab{debouncedSave=(0,e.debounce)(()=>this.plugin.saveSettings(),500,!0);constructor(e,t){super(e,t),this.app=e,this.plugin=t}display(){let{containerEl:e}=this;e.empty();for(let t of c)this.addGranularitySection(e,t)}addGranularitySection(t,r){let i=this.plugin.settings.granularities[r];t.createEl(`h3`,{text:va[r]}),new e.Setting(t).setName(`Enabled`).addToggle(e=>e.setValue(i.enabled).onChange(async e=>{this.plugin.settings.granularities[r].enabled=e,await this.plugin.saveSettings()}));let a=new e.Setting(t).setName(`Format`).setDesc(`Moment.js date format string`).addText(e=>{e.setPlaceholder(n[r]).setValue(i.format).onChange(async e=>{let t=g(e,r);a.descEl.setText(t||`Moment.js date format string`),a.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[r].format=e,this.debouncedSave()})}),o=new e.Setting(t).setName(`Folder`).addText(e=>{e.setValue(i.folder).onChange(async e=>{let t=_a(this.app,e);o.descEl.setText(t||``),o.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[r].folder=e,this.debouncedSave()}),new ha(this.app,e.inputEl)}),s=new e.Setting(t).setName(`Template`).addText(e=>{e.setValue(i.templatePath??``).onChange(async e=>{let t=ga(this.app,e);s.descEl.setText(t||``),s.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[r].templatePath=e||void 0,this.debouncedSave()}),new ma(this.app,e.inputEl)})}},ba={en:`en-gb`,zh:`zh-cn`,"zh-TW":`zh-tw`,ru:`ru`,ko:`ko`,it:`it`,id:`id`,ro:`ro`,"pt-BR":`pt-br`,cz:`cs`,da:`da`,de:`de`,es:`es`,fr:`fr`,no:`nn`,pl:`pl`,pt:`pt`,tr:`tr`,hi:`hi`,nl:`nl`,ar:`ar`,ja:`ja`};function xa(){let e=localStorage.getItem(`language`)||`en`,t=navigator.language?.toLowerCase(),n=ba[e]??e;t?.startsWith(e)&&(n=t);let r=window.moment.locale(n);console.debug(`[Periodic Notes] Configured locale: requested ${n}, got ${r}`)}var Sa=class extends e.Plugin{settings;ribbonEl;cache;async onload(){(0,e.addIcon)(`calendar-day`,ua),(0,e.addIcon)(`calendar-week`,da),(0,e.addIcon)(`calendar-month`,fa),(0,e.addIcon)(`calendar-year`,pa),await this.loadSettings(),xa(),this.ribbonEl=null,this.cache=new se(this.app,this),this.openPeriodicNote=this.openPeriodicNote.bind(this),this.addSettingTab(new ya(this.app,this)),this.configureRibbonIcons(),this.configureCommands(),this.registerView(o,e=>new ia(e,this)),this.addCommand({id:`show-calendar`,name:`Show calendar`,checkCallback:e=>{if(e)return this.app.workspace.getLeavesOfType(o).length===0;this.app.workspace.getRightLeaf(!1)?.setViewState({type:o})}})}configureRibbonIcons(){this.ribbonEl?.detach();let e=c.find(e=>this.settings.granularities[e].enabled);if(e){let t=aa[e];this.ribbonEl=this.addRibbonIcon(`calendar-${e}`,t.labelOpenPresent,t=>{t.type!==`auxclick`&&this.openPeriodicNote(e,window.moment(),{inNewSplit:Di(t)})}),this.ribbonEl.addEventListener(`contextmenu`,e=>{e.preventDefault(),la(this,{x:e.pageX,y:e.pageY})})}}configureCommands(){for(let e of c)ca(this.app,this,e).forEach(this.addCommand.bind(this))}async loadSettings(){let e=await this.loadData(),t=structuredClone(i);if(e?.granularities)for(let n of c)e.granularities[n]&&(t.granularities[n]={...t.granularities[n],...e.granularities[n]});this.settings=t}async saveSettings(){await this.saveData(this.settings),this.configureRibbonIcons(),this.app.workspace.trigger(`periodic-notes:settings-updated`)}async createPeriodicNote(e,t){let n=d(this.settings,e),r=l(this.settings,e),i=t.format(r),a=C(i,e,t,r,await ee(this.app,n.templatePath,e)),o=await ne(this.app,i,n);return this.app.vault.create(o,a)}getPeriodicNote(e,t){return this.cache.getPeriodicNote(e,t)}isPeriodic(e,t){return this.cache.isPeriodic(e,t)}findAdjacent(e,t){return this.cache.findAdjacent(e,t)}findInCache(e){return this.cache.find(e)}async openPeriodicNote(e,t,n){let{inNewSplit:r=!1}=n??{},{workspace:i}=this.app,a=this.cache.getPeriodicNote(e,t);a||=await this.createPeriodicNote(e,t),await(r?i.getLeaf(`split`):i.getLeaf()).openFile(a,{active:!0})}};module.exports=Sa;
+`,va=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getMarkdownFiles().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}},ya=class extends e.AbstractInputSuggest{onSelectCallback;constructor(e,t,n){super(e,t),this.onSelectCallback=n}getSuggestions(e){let t=e.toLowerCase();return this.app.vault.getAllFolders().filter(e=>e.path.toLowerCase().contains(t))}renderSuggestion(e,t){t.setText(e.path)}selectSuggestion(e){this.setValue(e.path),this.onSelectCallback?.(e.path),this.close()}};function ba(e,t){return t?e.metadataCache.getFirstLinkpathDest(t,``)?``:`Template file not found`:``}function xa(t,n){return!n||n===`/`||t.vault.getAbstractFileByPath((0,e.normalizePath)(n))?``:`Folder not found in vault`}var Sa={day:`Daily Notes`,week:`Weekly Notes`,month:`Monthly Notes`,year:`Yearly Notes`},Ca=class extends e.PluginSettingTab{debouncedSave=(0,e.debounce)(()=>this.plugin.saveSettings(),500,!0);constructor(e,t){super(e,t),this.app=e,this.plugin=t}display(){let{containerEl:e}=this;e.empty();for(let t of l)this.addGranularitySection(e,t)}addGranularitySection(t,n){let i=this.plugin.settings.granularities[n];t.createEl(`h3`,{text:Sa[n]}),new e.Setting(t).setName(`Enabled`).addToggle(e=>e.setValue(i.enabled).onChange(async e=>{this.plugin.settings.granularities[n].enabled=e,await this.plugin.saveSettings()}));let a=new e.Setting(t).setName(`Format`).setDesc(`Moment.js date format string`).addText(e=>{e.setPlaceholder(r[n]).setValue(i.format).onChange(async e=>{let t=_(e,n);a.descEl.setText(t||`Moment.js date format string`),a.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].format=e,this.debouncedSave()})}),o=new e.Setting(t).setName(`Folder`).addText(e=>{e.setValue(i.folder).onChange(async e=>{let t=xa(this.app,e);o.descEl.setText(t||``),o.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].folder=e,this.debouncedSave()}),new ya(this.app,e.inputEl)}),s=new e.Setting(t).setName(`Template`).addText(e=>{e.setValue(i.templatePath??``).onChange(async e=>{let t=ba(this.app,e);s.descEl.setText(t||``),s.descEl.toggleClass(`has-error`,!!t),this.plugin.settings.granularities[n].templatePath=e||void 0,this.debouncedSave()}),new va(this.app,e.inputEl)})}},wa={en:`en-gb`,zh:`zh-cn`,"zh-TW":`zh-tw`,ru:`ru`,ko:`ko`,it:`it`,id:`id`,ro:`ro`,"pt-BR":`pt-br`,cz:`cs`,da:`da`,de:`de`,es:`es`,fr:`fr`,no:`nn`,pl:`pl`,pt:`pt`,tr:`tr`,hi:`hi`,nl:`nl`,ar:`ar`,ja:`ja`};function Ta(){let e=localStorage.getItem(`language`)||`en`,t=navigator.language?.toLowerCase(),n=wa[e]??e;t?.startsWith(e)&&(n=t);let r=window.moment.locale(n);console.debug(`[Periodic Notes] Configured locale: requested ${n}, got ${r}`)}var Ea=class extends e.Plugin{settings;ribbonEl;cache;async onload(){(0,e.addIcon)(`calendar-day`,ma),(0,e.addIcon)(`calendar-week`,ha),(0,e.addIcon)(`calendar-month`,ga),(0,e.addIcon)(`calendar-year`,_a),await this.loadSettings(),Ta(),this.ribbonEl=null,this.cache=new de(this.app,this),this.openPeriodicNote=this.openPeriodicNote.bind(this),this.addSettingTab(new Ca(this.app,this)),this.configureRibbonIcons(),this.configureCommands(),this.registerView(s,e=>new ca(e,this)),this.addCommand({id:`show-calendar`,name:`Show calendar`,checkCallback:e=>{if(e)return this.app.workspace.getLeavesOfType(s).length===0;this.app.workspace.getRightLeaf(!1)?.setViewState({type:s})}})}configureRibbonIcons(){this.ribbonEl?.detach();let e=l.find(e=>this.settings.granularities[e].enabled);if(e){let t=la[e];this.ribbonEl=this.addRibbonIcon(`calendar-${e}`,t.labelOpenPresent,t=>{t.type!==`auxclick`&&this.openPeriodicNote(e,window.moment(),{inNewSplit:ji(t)})}),this.ribbonEl.addEventListener(`contextmenu`,e=>{e.preventDefault(),pa(this,{x:e.pageX,y:e.pageY})})}}configureCommands(){for(let e of l)fa(this.app,this,e).forEach(this.addCommand.bind(this))}async loadSettings(){let e=await this.loadData(),t=structuredClone(a);if(e?.granularities)for(let n of l)e.granularities[n]&&(t.granularities[n]={...t.granularities[n],...e.granularities[n]});this.settings=t}async saveSettings(){await this.saveData(this.settings),this.configureRibbonIcons(),this.app.workspace.trigger(`periodic-notes:settings-updated`)}async createPeriodicNote(e,t){let n=f(this.settings,e),r=u(this.settings,e),i=t.format(r),a=ie(i,e,t,r,await ae(this.app,n.templatePath,e)),o=await se(this.app,i,n);return this.app.vault.create(o,a)}getPeriodicNote(e,t){return this.cache.getPeriodicNote(e,t)}isPeriodic(e,t){return this.cache.isPeriodic(e,t)}findAdjacent(e,t){return this.cache.findAdjacent(e,t)}findInCache(e){return this.cache.find(e)}async openPeriodicNote(e,t,n){let{inNewSplit:r=!1}=n??{},{workspace:i}=this.app,a=this.cache.getPeriodicNote(e,t);a||=await this.createPeriodicNote(e,t),await(r?i.getLeaf(`split`):i.getLeaf()).openFile(a,{active:!0})}};module.exports=Ea;
