+++ src/template.test.ts
@@ -1,87 +1,6 @@
 import { describe, expect, test } from "bun:test";
 
-// applyTemplate can't be imported directly (template.ts imports obsidian).
-// Re-implement the pure template token replacement for testing.
-const WEEKDAYS = [
-  "sunday",
-  "monday",
-  "tuesday",
-  "wednesday",
-  "thursday",
-  "friday",
-  "saturday",
-] as const;
-
-function getDaysOfWeek(): string[] {
-  let weekStart = window.moment.localeData().firstDayOfWeek();
-  const daysOfWeek = [...WEEKDAYS];
-  while (weekStart) {
-    const day = daysOfWeek.shift();
-    if (day) daysOfWeek.push(day);
-    weekStart--;
-  }
-  return daysOfWeek;
-}
-
-function getDayOfWeekNumericalValue(name: string): number {
-  return Math.max(0, getDaysOfWeek().indexOf(name.toLowerCase()));
-}
-
-type Granularity = "day" | "week" | "month" | "year";
-
-function applyTemplate(
-  filename: string,
-  granularity: Granularity,
-  date: moment.Moment,
-  format: string,
-  rawContents: string,
-): string {
-  let contents = rawContents
-    .replace(/{{\s*date\s*}}/gi, filename)
-    .replace(/{{\s*time\s*}}/gi, window.moment().format("HH:mm"))
-    .replace(/{{\s*title\s*}}/gi, filename);
-
-  if (granularity === "day") {
-    contents = contents
-      .replace(
-        /{{\s*yesterday\s*}}/gi,
-        date.clone().subtract(1, "day").format(format),
-      )
-      .replace(/{{\s*tomorrow\s*}}/gi, date.clone().add(1, "d").format(format));
-  }
-
-  if (granularity === "week") {
-    contents = contents.replace(
-      new RegExp(`{{\\s*(${WEEKDAYS.join("|")})\\s*:(.*?)}}`, "gi"),
-      (_, dayOfWeek, momentFormat) => {
-        const day = getDayOfWeekNumericalValue(dayOfWeek);
-        return date.weekday(day).format(momentFormat.trim());
-      },
-    );
-  }
-
-  if (granularity === "month" || granularity === "year") {
-    const pattern = new RegExp(
-      `{{\\s*(${granularity})\\s*(([-+]\\d+)([ymwdhs]))?\\s*(:.+?)?}}`,
-      "gi",
-    );
-    contents = contents.replace(
-      pattern,
-      (_, _token, calc, timeDelta, unit, momentFormat) => {
-        const periodStart = date.clone().startOf(granularity);
-        if (calc) {
-          periodStart.add(parseInt(timeDelta, 10), unit);
-        }
-        if (momentFormat) {
-          return periodStart.format(momentFormat.substring(1).trim());
-        }
-        return periodStart.format(format);
-      },
-    );
-  }
-
-  return contents;
-}
+import { applyTemplate } from "./templateRender";
 
 describe("applyTemplate", () => {
   test("replaces date token", () => {
